% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplus_data.R
\name{add_time}
\alias{add_time}
\title{Add a padding time column for further aggregation.}
\usage{
add_time(data, based_col = NULL, interval, new_name = NULL,
  one_year = FALSE)
}
\arguments{
\item{data}{A data frame containing at least one datetime variable of class
'Date', class 'POSIXct' or class 'POSIXlt'.}

\item{based_col}{A character indicates the name of 'POSIXt' or 'Date' column.
If not given, it will be auto-detected.}

\item{interval}{The interval of the added datetime variable, which should be
higher than the interval of the input datetime variable. If 'NULL', it will
be one level higher than the interval of the input datetime variable. It can
also be an integer indicates the interval of minutes. In conclusion,
'interval' should be a number or one of c("year", "quarter", "month", "week",
"yday", "mday", "wday", "day", "date", "hour", "min", "second").}

\item{new_name}{The column name of the added variable. If 'NULL' it will be
the name of the original datetime variable with the interval name added to
it, separeted by an underscore.}

\item{one_year}{If TRUE, the (year+1)-01-01 will be replaced with
(year)-12-31.}
}
\description{
\code{add_time} adds a padding time column in a data.table for further
easy data aggregation. Basically, this is a wrapper function of
\code{\link[padr]{thicken}} but with an extra feature to give an exact number
of minutes of interval. It can also replace the datetime that is
(year+1)-01-01 with (year)-12-31.
}
