% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplus_data.R
\name{agg_by_time}
\alias{agg_by_time}
\title{Aggregate data according to given time interval.}
\usage{
agg_by_time(data, based_col = NULL, interval = "year", by_col = NULL,
  new_name = NULL, fun = "mean", one_year = FALSE)
}
\arguments{
\item{data}{A data frame containing at least one datetime variable of class
'Date', class 'POSIXct' or class 'POSIXlt'.}

\item{based_col}{A character indicates the name of 'POSIXt' or 'Date' column.
If not given, it will be auto-detected.}

\item{interval}{The interval of the added datetime variable, which should be
higher than the interval of the input datetime variable. If 'NULL', it will
be one level higher than the interval of the input datetime variable. It can
also be an integer indicates the interval of minutes. In conclusion,
'interval' should be a number or one of c("year", "quarter", "month", "week",
"yday", "mday", "wday", "day", "date", "hour", "min", "second").}

\item{by_col}{For grouping using the feature of \code{by} argument in
data.table.}

\item{new_name}{The column name of the added variable. If 'NULL' it will be
the name of the original datetime variable with the interval name added to
it, separeted by an underscore.}

\item{fun}{A character indicates the function to apply during aggregation.}

\item{one_year}{If TRUE, the (year+1)-01-01 will be replaced with
(year)-12-31.}
}
\description{
\code{agg_by_time} adds a padding time column in a data.table and aggregate
time series data according to given time interval.
}
