% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{annual_sch}
\alias{annual_sch}
\title{Create an annual schedule data for later used in EnergyPlus.}
\usage{
annual_sch(data, date_col = NULL, down = TRUE, leap_day = TRUE,
  rp_na = NA, tz = Sys.timezone())
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{date_col}{A character indicates the name of 'POSIXt' or 'Date' column.
If not given, it will be auto-detected.}

\item{down}{If TRUE, the time sequence will start from '01-01 00:00:00'
offsetted by inverval.}

\item{leap_day}{If FALSE, and the input year is a leap year, '02-29' related
data will be deleted.}

\item{rp_na}{A value to replace the missing values in the merged data.table.
Default is NA.}

\item{tz}{A character indicates the time zone of the time sequence. The
default value is the current system time zone.}

\item{interval}{An integer indicates the interval mins of the time sequence.}
}
\value{
It returns a time sequence of giving year and inverval.
}
\description{
\code{annual_sch} is a function to merge the input time series data.table
object with an annual time sequence. It returns a merged data.table of the
input and an annual time sequence with the same year of the input.
}
