% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{annual_seq}
\alias{annual_seq}
\title{Create an annual time sequence for data easily merging.}
\usage{
annual_seq(interval = 1L, year = current_year(), tz = Sys.timezone(),
  down = TRUE, leap_day = TRUE)
}
\arguments{
\item{interval}{An integer indicates the interval mins of the time sequence.}

\item{year}{An integer indicates the year of the time sequence. The default
value is the current year.}

\item{tz}{A character indicates the time zone of the time sequence. The
default value is the current system time zone.}

\item{down}{If TRUE, the time sequence will start from '01-01 00:00:00'
offsetted by inverval.}

\item{leap_day}{If FALSE, and the input year is a leap year, '02-29' related
data will be deleted.}
}
\value{
It returns a time sequence of giving year and inverval.
}
