% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl-epw.R
\name{as_EpwDate}
\alias{as_EpwDate}
\title{Convert to EnergyPlus Weather File date}
\usage{
as_EpwDate(x, leapyear = TRUE)
}
\arguments{
\item{x}{An integer vector or a character vector.}

\item{leapyear}{Whether support leap year. Default: \code{TRUE}}
}
\description{
\code{as_EpwDate()} converts inputs to EnergyPlus Weather File (EPW) dates.
}
\details{
EnergyPlus supports multiple formats of date specification
Reference: Table 2.14, Chap 2 Weather Converter Program, Auxiliary Program

Those formats include:
\enumerate{
\item Julian day of year
\item num_Month/num_Day
\item num_Month/num_Day/num_Year (only for DataPeriod)
\item num_Day alpha_Month
\item alpha_Month num_Day
\item num Weekday In Month (only for Holiday/DaylightSavingPeriod)
\item last Weekday In Month (only for Holiday/DaylightSavingPeriod)
}
}
\keyword{internal}
