% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bias}
\alias{bias}
\title{Calculate different percentage between two numeric vector.}
\usage{
bias(x, y)
}
\arguments{
\item{x, y}{A numeric vector to calculate with.}
}
\value{
The different percentage from x to y.
}
\description{
Calculate different percentage between two numeric vector.
}
