% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{case_cal}
\alias{case_cal}
\title{Apply a function between groups}
\usage{
case_cal(data, case_col, col_pattern = NULL, fun, case_order = TRUE,
  melt = FALSE)
}
\arguments{
\item{data}{A data.table object.}

\item{case_col}{Unquoted column name that is used as the group column.}

\item{col_pattern}{A RegEx indicates the column that the function applys to.}

\item{fun}{A character indicates the name of the function. It must be a
function that takes two arguments, i.e. cvrmse(x, y).}

\item{case_order}{If TRUE, the function applied will take the appearance
order in group column. For example, the group column contains three unique
c("a", "b", "c"), and only fun(a, b), fun(a, c) and fun(b, c) will be
calculated. If FALSE, all permutation without duplicated elements will be
calculated.}

\item{melt}{If TRUE, a molten data.table will be returned with the first
column named "case" showing what function and arguments has been applied.}
}
\value{
A data.table
}
\description{
\code{case_cal} .
}
