% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{col_names}
\alias{col_names}
\title{Column name extraction using RegEx.}
\usage{
col_names(data, pattern = ".*", regex = TRUE, ignore_case = TRUE,
  invert = FALSE)
}
\arguments{
\item{data}{A data frame that you want to extract column names from.}

\item{pattern}{A RegEx used for column name extraction. If multiple patterns
are given, then all items in the vector will be seperately used for
searching. Default is ".*", which will return all column names of the input
data frame.}

\item{regex}{If FALSE, \code{pattern} is treated as a character vector of
column names.}

\item{ignore_case}{A logical value. If FALSE, the pattern matching is case
sensitive and if TRUE, case is ignored during matching. Default is TRUE.}

\item{invert}{A logical value. If TRUE return column names that do not match.
Default is FALSE.}
}
\value{
A character vector of matched column names.
}
\description{
\code{col_names} returns all column names that are matched by the specifed
RegEx. Basically, it is a wrapper function of \code{\link[base]{grep}} but
can take multiple RegExs at a time.
}
\examples{
col_names(mtcars)
col_names(mtcars, "t$")
col_names(mtcars, c("^c", "t$"))
col_names(mtcars, "cyl", invert = TRUE)
col_names(mtcars, c("cly", "disp"), regex = FALSE)
}
