% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{current_year}
\alias{current_year}
\title{A function to return an integer of current calender year.}
\usage{
current_year()
}
\value{
An integer of current calender year.
}
\description{
A function to return an integer of current calender year.
}
\examples{
current_year()
}
