% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cvrmse}
\alias{cvrmse}
\title{Calculate CV(RMSE) between two numeric vector.}
\usage{
cvrmse(x, y, na_rm = TRUE, ...)
}
\arguments{
\item{x, y}{A numeric vector to calculate with.}

\item{na_rm}{If TRUE, NA values will be removed.}
}
\value{
The CV(RMSE) of x to y.
}
\description{
Calculate CV(RMSE) between two numeric vector.
}
