% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_melt}
\alias{data_melt}
\title{Melt a data.frame with column selection using RegEx.}
\usage{
data_melt(data, id_pattern, measure_pattern, ignore_case = TRUE,
  variable_name = "variable", value_name = "value", na_rm = FALSE,
  variable_factor = TRUE, value_factor = FALSE)
}
\arguments{
\item{data}{A data.frame or data.table object to manipulate.}

\item{id_pattern}{A RegEx used for id column selection.}

\item{measure_pattern}{A RegEx used for measure column selection.}

\item{variable_name}{Name for the measured variable names column. The default
name is 'variable'.}

\item{value_name}{Name for the molten data values column. The default name is
'value'.}

\item{na_rm}{If TRUE, NA values will be removed from the molten data.}

\item{variable_factor}{If TRUE, the variable column will be converted to
factor, else it will be a character column.}

\item{value_factor}{If TRUE, the value column will be converted to factor,
else the molten value type is left unchanged.}
}
\value{
A molten data.table.
}
\description{
Basically, \code{data_melt} is a wrapper function of
\code{\link{data.table::melt.data.table}} but enables to use RegEx for
convenient selection of both id columns and measure columns.
}
\examples{
data_melt(iris, id_pattern = "^S")
}
