% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{datetime}
\alias{datetime}
\alias{time_from_eplus}
\alias{datetime}
\alias{time_to_eplus}
\alias{datetime}
\alias{yhour}
\alias{datetime}
\alias{add_time}
\title{Transform EnergyPlus DateTime format into POSIXct class.}
\usage{
time_from_eplus(x, year = current_year(), tz = Sys.timezone())

time_to_eplus(x)

yhour(x, one_year = FALSE, no_leap = FALSE)

add_time(data, base = NULL, new = NULL, step, toward = c("up", "down",
  "center"), no_leap = FALSE)
}
\arguments{
\item{x}{An EnergyPlus DateTime format string vector.}

\item{year}{Year number.}

\item{tz}{Time zone of the result POSIXct sequence.}

\item{one_year}{If TRUE, the calculation will based on the earliest year in
the input}

\item{no_leap}{If TRUE, and the year is a leap year, Feb 29th will be
neglected during the calculation.}

\item{data}{A data frame containing at least one datetime column.}

\item{base}{A character indicates the name of datetime column.  If not given,
it will be auto-detected.}

\item{new}{The column name of the added variable. If 'NULL' it will be
the name of the original datetime variable with the interval name added to
it, separeted by an underscore.}

\item{step}{The timestep of the added datetime variable, which should be
higher than the interval of the input datetime variable. If 'NULL', it will
be one level higher than the interval of the input datetime variable. It can
also be an integer indicates the interval of minutes. In conclusion,
'interval' should be a number or one of c("year", "quarter", "month", "week",
"yday", "mday", "wday", "day", "date", "hour", "min", "second").}

\item{x}{A POSIXct object.}

\item{x}{A datetime object.}

\item{one_year}{If TRUE, the (year+1)-01-01 will be replaced with
(year)-12-31.}
}
\value{
A POSIXct sequence.

A string vector

A numeric vector indicates the hour number in the year.
}
\description{
\code{time_from_eplus} takes an EnergyPlus DateTime format string, e.g.
" %m/%d  %H:%M:%d" into a POSIXct sequence in R.

\code{time_to_eplus} takes a POSIXct and formats it into a string with
EnergyPlus DateTime format, e.g." %m/%d  %H:%M:%d".

\code{yhour} returns the hour number in a year.

\code{add_time} adds a padding time column in a data.frame for further easy
data aggregation. It can also replace the datetime that is (year+1)-01-01
with (year)-12-31.
}
