% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idd.R
\name{download_idd}
\alias{download_idd}
\title{Download EnergyPlus Input Data Dictionary (IDD) File}
\usage{
download_idd(ver = "latest", dir = getwd())
}
\arguments{
\item{ver}{The EnergyPlus version number, e.g., 8.7, "8.7" or "8.7.0". The
special value \code{"latest"} means the latest version (fetched from GitHub
releases).}

\item{dir}{A directory to indicate where to save the IDD file. Default:
current working directory.}
}
\value{
A full path of the downloaded IDD file.
}
\description{
\code{download_idd} downloads EnergyPlus Input Data Dictionary (IDD) file from
\href{https://github.com/NREL/EnergyPlus}{EnergyPlus GitHub Repository}. It is
useful in case where you only want to edit an EnergyPlus Input Data File
(IDF) directly but do not want to install whole EnergyPlus software.
}
\examples{
\dontrun{
download_idd()
download_idd(8.8)
}
}
\seealso{
\code{\link[=use_idd]{use_idd()}}
}
