% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl-idf.R
\name{dup_idf_object}
\alias{dup_idf_object}
\title{Duplicate existing objects}
\usage{
dup_idf_object(
  idd_env,
  idf_env,
  dt_object,
  level = eplusr_option("validate_level")
)
}
\arguments{
\item{idd_env}{An environment or list contains IDD tables including class,
field, and reference.}

\item{idf_env}{An environment or list contains IDF tables including object,
value, and reference.}

\item{dt_object}{A \code{\link[data.table:data.table]{data.table::data.table()}} that contains object data.
Usually created using \code{\link[=expand_idf_dots_value]{expand_idf_dots_value()}}.}

\item{level}{Validate level. Default: \code{eplusr_option("validate_level")}.}
}
\value{
The duplicated object data in a named list of 3
\code{\link[data.table:data.table]{data.table::data.table()}}s, i.e. \code{object}, \code{value} and \code{reference}.
}
\description{
Duplicate existing objects
}
\keyword{internal}
