% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epat.R
\name{edit_epat}
\alias{edit_epat}
\title{A Shiny Gadget for editing EPAT job}
\usage{
edit_epat(epat, parse = FALSE)
}
\arguments{
\item{json}{A EPAT json file or json string.}
}
\value{
A eplusr job with "job_type" attribute of "epat".
}
\description{
\code{edit_pat} takes an EPAT json project file, starts a Shiny Gadget to
edit it interactively, and returns an EPAT job that can be directly called by
\link[eplusr]{run_job}.
}
