% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.R
\name{empty_idf}
\alias{empty_idf}
\title{Create an Empty Idf}
\usage{
empty_idf(ver = "latest")
}
\arguments{
\item{ver}{Any acceptable input of \code{\link[=use_idd]{use_idd()}}. If \code{latest}, which is the
default, the latest IDD released version is used.}
}
\value{
An \link{Idf} object
}
\description{
\code{empty_idf()} takes a valid IDD version and creates an empty \link{Idf} object
that only contains a Version object.
}
\examples{
if (is_avail_idd(8.8)) empty_idf(8.8)
}
