% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplus_app.R
\name{epat}
\alias{epat}
\title{A Shiny app for EnergyPlus Parametric Analysis in R.}
\usage{
epat()
}
\description{
\code{epat} will launch a shiny app called EPAT (EnergyPlus Parametric
Analysis Toolkit).
}
\examples{
epat()
}
