% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{eplus_time_fmt}
\alias{eplus_time_fmt}
\title{Format POSIXct time into EnergyPlus Date/Time format.}
\usage{
eplus_time_fmt(data, date_col = NULL, std_fmt = TRUE, copy = TRUE)
}
\arguments{
\item{data}{A data.table object to manipulate.}

\item{date_col}{A character indicates the name of 'POSIXt' or 'Date' column.
If not given, it will be auto-detected.}

\item{std_fmt}{If FALSE, '00:00' will be replaced with EnergyPlus-flavored
'24:00' format.}

\item{copy}{Whether hard copy the input to protect it unchanged.}
}
\description{
\code{eplus_time_fmt}: A function to transform "00:00" to "24:00" and format
date into EnergyPlus format, i.e. "mm/dd HH:MM:SS". Use with caution because
the 'POSIXct' and 'Date' data will be converted into characters.
}
