% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{eplus_time_trans}
\alias{eplus_time_trans}
\title{Transform EnergyPlus Date/Time format into POSIXct format.}
\usage{
eplus_time_trans(data, year = current_year(), eplus_date_col = "Date/Time",
  new_date_col = "datetime", tz = Sys.timezone(), keep_ori = FALSE,
  copy = FALSE)
}
\arguments{
\item{data}{A data.table object to manipulate.}

\item{year}{An integer indicates the year of the time sequence. The default
value is the current year.}

\item{eplus_date_col}{A character indicates the column name of EnergyPlus
time in the input.}

\item{new_date_col}{A character indicates the name of the new transformed
POSIXt column.}

\item{tz}{A character indicates the time zone of the transformed time column.
The default value is the current system time zone.}

\item{keep_ori}{Whether to keep the original EnergyPlus time column. Default
is FALSE which means the EnergyPlus time column will be deleted.}

\item{copy}{Whether hard copy the input to protect it unchanged. Default is
FALSE.}
}
\value{
A data.table with first column being the transformed POSIXct column
from EnergyPlus Date/Time format column.
}
\description{
\code{eplus_time_trans} takes a data.table as input, and transform the column
with EnergyPlus Date/Time format into a column of POSIXct.
}
