% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{eplusr-deprecated}
\alias{eplusr-deprecated}
\alias{with_speed}
\title{Deprecated functions since eplusr v0.15.0}
\usage{
with_speed(expr)
}
\arguments{
\item{expr}{An expression to be evaluated.}
}
\description{
Deprecated functions since eplusr v0.15.0
}
