% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl-idf.R
\name{expand_idf_regex}
\alias{expand_idf_regex}
\title{Parse regular expression of object field values}
\usage{
expand_idf_regex(
  idd_env,
  idf_env,
  pattern,
  replacement = NULL,
  class = NULL,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{idd_env}{An environment or list contains IDD tables including class,
field, and reference.}

\item{idf_env}{An environment or list contains IDF tables including object,
value, and reference.}

\item{pattern}{A single string of regular expression used to match field
values}

\item{class}{A character vector specifying the target class names}

\item{pattern, ignore.case, perl, fixed, useBytes}{All of them are
directly passed to \link[base:grep]{base::grepl} and
\link[base:grep]{base::gsub} with the same default values.}
}
\value{
A named list of 2 \code{\link[data.table:data.table]{data.table::data.table()}}: \code{object} and \code{value}.
}
\description{
Parse regular expression of object field values
}
\keyword{internal}
