% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{find_eplus}
\alias{find_eplus}
\title{Find EnergyPlus locations on current computer.}
\usage{
find_eplus(ver = NULL, verbose = FALSE)
}
\arguments{
\item{ver}{The version to locate for. If NULL, all versions will be returned.}

\item{verbose}{If TRUE, extra message will be shown.}
}
\value{
A tibble object with two columns named 'path' and 'version'.
}
\description{
\code{find_eplus} try to find the locations of EnergyPlus on current
computer.
}
\examples{
find_eplus()
}
