% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idd.R
\name{format.Idd}
\alias{format.Idd}
\title{Format an Idd}
\usage{
\method{format}{Idd}(x, ...)
}
\arguments{
\item{x}{An \link{Idd} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A single length character vector.
}
\description{
Format an \link{Idd} into a string.
}
\examples{
\dontrun{
format(use_idd(8.8, download = "auto"))
}

}
