% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iddobj.R
\name{format.IddObject}
\alias{format.IddObject}
\title{Format an IddObject}
\usage{
\method{format}{IddObject}(x, ver = TRUE, ...)
}
\arguments{
\item{x}{An \link{IddObject} object.}

\item{ver}{If \code{TRUE}, a suffix of version string is added. Default: \code{TRUE}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A single length character vector.
}
\description{
Format an \link{IddObject} into a string. It is formatted the same way as
\code{IddObject$print(brief = TRUE)} but with a suffix of current IDD version.
}
\examples{
\dontrun{
format(use_idd(8.8, download = "auto")$Material)
}

}
