% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl-idd.R
\name{get_idd_class}
\alias{get_idd_class}
\title{Get class data}
\usage{
get_idd_class(idd_env, class = NULL, property = NULL, underscore = FALSE)
}
\arguments{
\item{idd_env}{An environment or list contains IDD tables including class,
field, and reference.}

\item{class}{An integer vector of valid class indexes or a character vector
of valid class names. If \code{NULL}, all classes are returned.}

\item{property}{A character vector of column names in class table to return.
If \code{NULL}, only class index columns are returned, plus column \code{rleid}.}

\item{underscore}{If \code{TRUE}, input class name will be converted into
underscore style name first and column \code{class_name_us} will be used
for matching.}
}
\value{
A data.table containing specified columns.
}
\description{
Get class data
}
\keyword{internal}
