% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl-idf.R
\name{get_idf_object}
\alias{get_idf_object}
\title{Get object data}
\usage{
get_idf_object(
  idd_env,
  idf_env,
  class = NULL,
  object = NULL,
  property = NULL,
  underscore = FALSE,
  ignore_case = FALSE
)
}
\arguments{
\item{idd_env}{An environment or list contains IDD tables including class,
field, and reference.}

\item{idf_env}{An environment or list contains IDF tables including object,
value, and reference.}

\item{class}{An integer vector of valid class indexes or a character vector
of valid class names. Default: \code{NULL}.}

\item{object}{An integer vector of valid object IDs or a character vector
of valid object names. Default: \code{NULL}.}

\item{property}{A character vector of column names in class table to return.}

\item{underscore}{If \code{TRUE}, input class name will be converted into
underscore style name first and column \code{class_name_us} will be used
for matching. Default: \code{FALSE}.}

\item{ignore_case}{If \code{TRUE}, input object name will be converted into lower
case and column \code{object_name_lower} will be used for matching.
converted into underscore style name first and column \code{class_name_us}
and \code{field_name_us} will be used for matching. Default: \code{FALSE}.}
}
\value{
A data.table.
}
\description{
Get object data
}
\keyword{internal}
