% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idd_object.R
\name{idd_object}
\alias{idd_object}
\title{Create an \code{IddObject} object.}
\usage{
idd_object(parent, class)
}
\arguments{
\item{parent}{An \link{Idd} object or a valid input for \code{\link[=use_idd]{use_idd()}}.}

\item{class}{A valid class name (a string).}
}
\value{
An \link{IddObject} object.
}
\description{
\code{idd_object()} takes a parent \code{Idd} object, a class name, and returns a
corresponding \link{IddObject}. For details, see \link{IddObject}.
}
\examples{
idd <- use_idd(8.8, download = "auto")

# get an IddObject using class name
idd_object(idd, "Material")
idd_object(8.8, "Material")
}
