% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.R
\docType{class}
\name{IDF}
\alias{IDF}
\title{Parse an EnergyPlus IDF file}
\value{
An IDF object
}
\description{
\code{IDF} R6 class is the core class that enable fully access to and
manipulate EnergyPlus IDF models.
}
\section{Usage}{
\preformatted{
idf <- IDF$new(path, idd)

idf$version()
idf$id_of_class(class)
idf$class_of_id(id)
idf$all_class(scope = "idf")
idf$all_id()

idf$is_valid_class(class, scope = "idf")
idf$is_valid_id(id)

idf$object(id)
idf$objects(id)
idf$objects_in_class(class)
idf$objects_in_group(group)

idf$out(format = c("asis", "sorted", "ori_bot", "ori_top"), comment = TRUE,
in_ip = FALSE)

idf$print()

print(idf)

}
}

\section{Arguments}{

\itemize{
\item \code{path}: Path to an EnergyPlus Input Data Dictionary (IDD) file, usually
named as \code{Energy+.idd}.
\item \code{group}: A valid group name or valid group names.
\item \code{class}: A valid class name or valid class names.
}
}

\section{Detail}{


\code{IDD$new()} parses an EnergyPlus Input Data Dictionary (IDD) file, and
returns an IDD object.

\code{$version()} returns the version string of current idd file.

\code{$build()} returns the build tag string of current idd file.

\code{$group_name(class)} returns group name that that \code{class} belong to.

\code{$class_name(group)} returns class names of that \code{group}. If \code{group} not
given, all class names in current IDD are returned.

\code{$group_order(group)} returns integer orders (orders of name apperarance in
the IDD file) of that \code{group}.

\code{$class_order(class)} returns integer orders (orders of name apperarance in
the IDD file) of that \code{class}.

\code{$orders()} returns the all order data (stored as a data.table with four
columns, i.e. "group", "group_order", "class", "class_order").

\code{$object(class)} returns an IDDObject of that \code{class}.

\code{$objects(class)} returns a list of IDDObjects of \code{class}es. If \code{class} is
NULL, all IDDObjects in current IDD are returned.

\code{$objects_in_group(group)} returns a list of IDDObjects in that \code{group}.

\code{$is_valid_group(group)} return \code{TRUE} if the input is a valid \code{group} name.

\code{$is_valid_class(class)} return \code{TRUE} if the input is a valid \code{class} name.
}

\references{
\href{https://github.com/NREL/EnergyPlus/tree/develop/src/IDF_Editor}{IDFEditor
source code}
}
\author{
Hongyuan Jia
}
