% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\docType{class}
\name{IdfGeometry}
\alias{IdfGeometry}
\alias{idf_geometry}
\title{Modify and Visualize an EnergyPlus Model Geometry}
\usage{
idf_geometry(parent, object = NULL)
}
\description{
\code{IdfGeometry} is an abstraction of a collection of geometry in an \link{Idf}. It
provides more detail methods to query geometry properties, update geometry
vertices and visualize geometry in 3D using the
\href{https://cran.r-project.org/package=rgl}{rgl} package.
}
\examples{

## ------------------------------------------------
## Method `IdfGeometry$new`
## ------------------------------------------------

\dontrun{
# example model shipped with eplusr from EnergyPlus v8.8
path_idf <- system.file("extdata/1ZoneUncontrolled.idf", package = "eplusr") # v8.8

# create from an Idf object
idf <- read_idf(path_idf, use_idd(8.8, "auto"))
geom <- idf$geometry()
geom <- IdfGeometry$new(idf)

# create from an IDF file
geom <- idf_geometry(path_idf)
geom <- IdfGeometry$new(path_idf)
}

## ------------------------------------------------
## Method `IdfGeometry$parent`
## ------------------------------------------------

geom$parent()

## ------------------------------------------------
## Method `IdfGeometry$rules`
## ------------------------------------------------

geom$rules()

## ------------------------------------------------
## Method `IdfGeometry$convert`
## ------------------------------------------------

\dontrun{
geom$convert()
}

## ------------------------------------------------
## Method `IdfGeometry$round_digits`
## ------------------------------------------------

\dontrun{
geom$round_digits()
}

## ------------------------------------------------
## Method `IdfGeometry$area`
## ------------------------------------------------

\dontrun{
geom$area()
}

## ------------------------------------------------
## Method `IdfGeometry$outward_normal`
## ------------------------------------------------

\dontrun{
geom$outward_normal()
}

## ------------------------------------------------
## Method `IdfGeometry$centroid`
## ------------------------------------------------

\dontrun{
geom$centroid()
}

## ------------------------------------------------
## Method `IdfGeometry$azimuth`
## ------------------------------------------------

\dontrun{
geom$azimuth()
}

## ------------------------------------------------
## Method `IdfGeometry$tilt`
## ------------------------------------------------

\dontrun{
geom$tilt()
}

## ------------------------------------------------
## Method `IdfGeometry$extract`
## ------------------------------------------------

\dontrun{
geom$extract()
}

## ------------------------------------------------
## Method `IdfGeometry$view`
## ------------------------------------------------

\dontrun{
idf$view()
idf$view(render_by = "zone")
idf$view(new, render_by = "construction")
}

## ------------------------------------------------
## Method `IdfGeometry$save_snapshot`
## ------------------------------------------------

\dontrun{
geom$view()
geom$save_snapshot(tempfile(fileext = ".png"))
}


## ------------------------------------------------
## Method `IdfGeometry$print`
## ------------------------------------------------

\dontrun{
geom$print()
}
}
\seealso{
\link{Idf} class
}
\author{
Hongyuan Jia
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IdfGeometry$new()}}
\item \href{#method-parent}{\code{IdfGeometry$parent()}}
\item \href{#method-rules}{\code{IdfGeometry$rules()}}
\item \href{#method-convert}{\code{IdfGeometry$convert()}}
\item \href{#method-round_digits}{\code{IdfGeometry$round_digits()}}
\item \href{#method-area}{\code{IdfGeometry$area()}}
\item \href{#method-outward_normal}{\code{IdfGeometry$outward_normal()}}
\item \href{#method-centroid}{\code{IdfGeometry$centroid()}}
\item \href{#method-azimuth}{\code{IdfGeometry$azimuth()}}
\item \href{#method-tilt}{\code{IdfGeometry$tilt()}}
\item \href{#method-extract}{\code{IdfGeometry$extract()}}
\item \href{#method-view}{\code{IdfGeometry$view()}}
\item \href{#method-save_snapshot}{\code{IdfGeometry$save_snapshot()}}
\item \href{#method-print}{\code{IdfGeometry$print()}}
\item \href{#method-clone}{\code{IdfGeometry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an \code{IdfGeometry} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$new(parent, object = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent}}{A path to an IDF file or an \link{Idf} object.}

\item{\code{object}}{A character vector of valid names or an integer
vector of valid IDs of objects to extract. If \code{NULL}, all
objects in geometry classes will be extracted.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{IdfGeometry} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# example model shipped with eplusr from EnergyPlus v8.8
path_idf <- system.file("extdata/1ZoneUncontrolled.idf", package = "eplusr") # v8.8

# create from an Idf object
idf <- read_idf(path_idf, use_idd(8.8, "auto"))
geom <- idf$geometry()
geom <- IdfGeometry$new(idf)

# create from an IDF file
geom <- idf_geometry(path_idf)
geom <- IdfGeometry$new(path_idf)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parent"></a>}}
\if{latex}{\out{\hypertarget{method-parent}{}}}
\subsection{Method \code{parent()}}{
Get parent \link{Idf} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$parent()}\if{html}{\out{</div>}}
}

\subsection{Details}{
\verb{$parent()} returns the parent \link{Idf} object of current \code{IdfGeometry}
object.
}

\subsection{Returns}{
An \link{Idf} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{geom$parent()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rules"></a>}}
\if{latex}{\out{\hypertarget{method-rules}{}}}
\subsection{Method \code{rules()}}{
Get global geometry rules
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$rules()}\if{html}{\out{</div>}}
}

\subsection{Details}{
\verb{$rules()} set global geometry rules.
}

\subsection{Returns}{
An \link{Idf} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{geom$rules()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convert"></a>}}
\if{latex}{\out{\hypertarget{method-convert}{}}}
\subsection{Method \code{convert()}}{
Convert simple geometry objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$convert(
  type = c("surface", "subsurface", "shading"),
  object = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A character vector giving what types of simplified
geometries should be converted. Should be a subset of
\code{"surface"}, \code{"subsurface"} and \code{"shading"}. Default is set to
all of them.}

\item{\code{object}}{A character vector of valid names or an integer
vector of valid IDs of targeting objects to be converted.
Default: \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
EnergyPlus provides several classes that allow for simplified entry
of geometries, such as \code{Wall:Exterior}, \code{Window} and etc.
\verb{$convert()} will generate detailed vertices from simplified geometry
specifications and replace the original object with its corresponding
detailed class, including:
\itemize{
\item \code{BuildingSurface:Detailed}
\item \code{FenestrationSurface:Detailed}
\item \code{Shading:Site:Detailed}
\item \code{Shading:Building:Detailed}
\item \code{Shading:Zone:Detailed}
}
}

\subsection{Returns}{
The modified \link{Idf} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
geom$convert()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-round_digits"></a>}}
\if{latex}{\out{\hypertarget{method-round_digits}{}}}
\subsection{Method \code{round_digits()}}{
Round digits on geometry vertices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$round_digits(digits = 4L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{An integer giving the number of decimal places to be
used. Default: \code{5}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$round_digits()} performs number rounding on geometry object
vertices. It may be useful for clean up IDF files generated using
OpenStudio which often gives vertices with long trailing digits.
}

\subsection{Returns}{
The modified \link{Idf} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
geom$round_digits()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-area"></a>}}
\if{latex}{\out{\hypertarget{method-area}{}}}
\subsection{Method \code{area()}}{
Get area
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$area(class = NULL, object = NULL, net = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{A character vector of valid geometry class names.
Default: \code{NULL}.}

\item{\code{object}}{A character vector of valid names or an integer
vector of valid IDs of targeting objects.
Default: \code{NULL}.}

\item{\code{net}}{If \code{TRUE}, the gross area is returned. If \code{FALSE}, the net
area is returned. Default: \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$area()} returns the area of surfaces in square meters.
}

\subsection{Returns}{
A \code{\link[data.table:data.table]{data.table::data.table()}} of 4 columns:
\itemize{
\item \code{id}: Integer type. Object IDs.
\item \code{name}: Character type. Object names.
\item \code{class}: Character type. Class names.
\item \code{area}: Numeric type. Surface Area in m2.
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
geom$area()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-outward_normal"></a>}}
\if{latex}{\out{\hypertarget{method-outward_normal}{}}}
\subsection{Method \code{outward_normal()}}{
Get outward normal vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$outward_normal(class = NULL, object = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{A character vector of valid geometry class names.
Default: \code{NULL}.}

\item{\code{object}}{A character vector of valid names or an integer
vector of valid IDs of targeting objects.
Default: \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$outward_normal()} returns the outward normal vector of surfaces.
}

\subsection{Returns}{
A \code{\link[data.table:data.table]{data.table::data.table()}} of 6 columns:
\itemize{
\item \code{id}: Integer type. Object IDs.
\item \code{name}: Character type. Object names.
\item \code{class}: Character type. Class names.
\item \code{x}: Numeric type. X axis value.
\item \code{y}: Numeric type. Y axis value.
\item \code{z}: Numeric type. Z axis value.
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
geom$outward_normal()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-centroid"></a>}}
\if{latex}{\out{\hypertarget{method-centroid}{}}}
\subsection{Method \code{centroid()}}{
Get centroid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$centroid(class = NULL, object = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{A character vector of valid geometry class names.
Default: \code{NULL}.}

\item{\code{object}}{A character vector of valid names or an integer
vector of valid IDs of targeting objects.
Default: \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$centroid()} returns the centroid of surfaces.
}

\subsection{Returns}{
A \code{\link[data.table:data.table]{data.table::data.table()}} of 6 columns:
\itemize{
\item \code{id}: Integer type. Object IDs.
\item \code{name}: Character type. Object names.
\item \code{class}: Character type. Class names.
\item \code{x}: Numeric type. X axis value.
\item \code{y}: Numeric type. Y axis value.
\item \code{z}: Numeric type. Z axis value.
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
geom$centroid()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-azimuth"></a>}}
\if{latex}{\out{\hypertarget{method-azimuth}{}}}
\subsection{Method \code{azimuth()}}{
Get azimuth
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$azimuth(class = NULL, object = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{A character vector of valid geometry class names.
Default: \code{NULL}.}

\item{\code{object}}{A character vector of valid names or an integer
vector of valid IDs of targeting objects.
Default: \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$azimuth()} returns the azimuth of surfaces in degree.
}

\subsection{Returns}{
A \code{\link[data.table:data.table]{data.table::data.table()}} of 4 columns:
\itemize{
\item \code{id}: Integer type. Object IDs.
\item \code{name}: Character type. Object names.
\item \code{class}: Character type. Class names.
\item \code{azimuth}: Numeric type. Azimuth in degree.
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
geom$azimuth()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tilt"></a>}}
\if{latex}{\out{\hypertarget{method-tilt}{}}}
\subsection{Method \code{tilt()}}{
Get tilt
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$tilt(class = NULL, object = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{A character vector of valid geometry class names.
Default: \code{NULL}.}

\item{\code{object}}{A character vector of valid names or an integer
vector of valid IDs of targeting objects.
Default: \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$tilt()} returns the tilt of surfaces in degree.
}

\subsection{Returns}{
A \code{\link[data.table:data.table]{data.table::data.table()}} of 4 columns:
\itemize{
\item \code{id}: Integer type. Object IDs.
\item \code{name}: Character type. Object names.
\item \code{class}: Character type. Class names.
\item \code{tilt}: Numeric type. Azimuth in degree.
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
geom$tilt()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract"></a>}}
\if{latex}{\out{\hypertarget{method-extract}{}}}
\subsection{Method \code{extract()}}{
Extract geometry data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$extract(type = NULL, class = NULL, object = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A character vector of valid geometry type names,
including:
\itemize{
\item \code{"surface"}
\item \code{"subsurface"}
\item \code{"shading"}
\item \code{"daylighting_point"} (daylighting control reference points)
}

If \code{NULL}, all types will be returned.}

\item{\code{class}}{A character vector of valid geometry class names.
Default: \code{NULL}.}

\item{\code{object}}{A character vector of valid names or an integer
vector of valid IDs of targeting objects.
Default: \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$extract()} returns a list that contains data of geometry objects
specified.
}

\subsection{Returns}{
A list of 4 columns:
\itemize{
\item \code{id}: Integer type. Object IDs.
\item \code{name}: Character type. Object names.
\item \code{class}: Character type. Class names.
\item \code{x}: Numeric type. X axis value.
\item \code{y}: Numeric type. Y axis value.
\item \code{z}: Numeric type. Z axis value.
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
geom$extract()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-view"></a>}}
\if{latex}{\out{\hypertarget{method-view}{}}}
\subsection{Method \code{view()}}{
View 3D geometry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$view(
  new = TRUE,
  clear = TRUE,
  show = "all",
  render_by = "surface_type",
  zone = NULL,
  surface = NULL,
  wireframe = TRUE,
  x_ray = FALSE,
  axis = TRUE,
  line_width = 1.5,
  line_color = "black",
  theta = 0,
  phi = -60,
  fov = 60,
  zoom = 1,
  background = "white",
  size = c(0, 30, 800)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new}}{If \code{TRUE}, a new rgl window will be open using
\code{\link[rgl:rgl.open]{rgl::rgl.open()}}. If \code{FALSE}, existing rgl window will be
reused if possible. Default: \code{FALSE}.}

\item{\code{render_by}}{A single string specifying the way of rendering the
geometry. Possible values are:\preformatted{* `"surface_type"`: Default. Render the model by surface type
  model. Walls, roofs, windows, doors, floors, and shading
  surfaces will have unqiue colors.
* `"boundary"`: Render the model by outside boundary condition.
  Only surfaces that have boundary conditions will be rendered
  with a color. All other surfaces will be white.
* `"construction"`: Render the model by surface constructions.
* `"zone"`: Render the model by zones assigned.
* `"normal"`: Render the model by surface normal. The outside
  face of a heat transfer face will be rendered as white and the
  inside face will be rendered as red.
}}

\item{\code{wireframe}}{If \code{TRUE}, the wireframe of each surface will be
shown. Default: \code{TRUE}.}

\item{\code{x_ray}}{If \code{TRUE}, all surfaces wll be rendered translucently.
Default: \code{FALSE}.}

\item{\code{axis}}{If \code{TRUE}, the X, Y and Z axes will be drawn at the
global origin. Default: \code{TRUE}.}

\item{\code{line_width}}{The line width of wireframes. Default: \code{1.5}.}

\item{\code{line_color}}{The color of wireframes. Default: \code{black}.}

\item{\code{theta}}{The rotation around z-axis in degrees. Default: \code{0}.}

\item{\code{phi}}{The azimuth angle in degrees. Default: \code{-60}.}

\item{\code{fov}}{The field-of-view angle in degrees. \code{0} means isometric.
Default: \code{60}.}

\item{\code{zoom}}{The zoom factor. Default: \code{1}.}

\item{\code{background}}{The color of the backports. Default: \code{"white"}.}

\item{\code{size}}{A numeric vector specifying the x-y coordiates, width and
height of the rgl device. The x-y coordinates are based on the
original located in the upper left corner of current screen.
By default, x-y coordinates are set to zeros.
\itemize{
\item If 1 number, the width and height of rgl windows will be
the same as input number.
\item If 2 numbers, the width and height of rgl windows will be
set based on input number
\item If 3 numbers, the first 2 numbers will be used as x-y
coordinates and the 3rd number will be used for both window
width and height.
\item If 4 numbers, the first 2 numbers will be used as x-y
coordinates. The next 2 specify the window size.
}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$view()} uses the \href{https://cran.r-project.org/package=rgl}{rgl}
package to visualize the IDF geometry in 3D in a similar way as
\href{https://www.openstudio.net}{OpenStudio}.

In the rgl window, you can control the view using your mouse:
\itemize{
\item Left button: Trackball
\item Right button: Pan
\item Middle button: Field-of-view (FOV). '0' means orthographic
projection.
\item Wheel: Zoom
}
}

\subsection{Returns}{
An \link{IdfViewer} object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
idf$view()
idf$view(render_by = "zone")
idf$view(new, render_by = "construction")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_snapshot"></a>}}
\if{latex}{\out{\hypertarget{method-save_snapshot}{}}}
\subsection{Method \code{save_snapshot()}}{
Capture and save current rgl view as an image
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$save_snapshot(filename, bring_to_front = TRUE, axis = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{A single string specifying the file name. Current
supported formats are \code{png}, \code{pdf}, \code{svg}, \code{ps}, \code{eps}, \code{tex}
and \code{pgf}.}

\item{\code{bring_to_front}}{If \code{TRUE}, The rgl window will be brought to
the front when saving. Default: \code{FALSE}.}

\item{\code{axis}}{If \code{TRUE}, the X, Y and Z axes will be saved in the
image. Default: \code{FALSE}.}

\item{\code{autoview}}{If \code{TRUE}, a new view will be created if there is no
existing one using
\href{../../eplusr/html/Idf.html#method-view}{\code{$view()}}
Default: \code{FALSE}.}

\item{\code{autoclose}}{If \code{TRUE}, current rgl window will be closed after
saving. Default: \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$save_view()} captures the current rgl view and saves it as an
image file to disk.
}

\subsection{Returns}{
The \code{Idf} object itself, invisibly.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
geom$view()
geom$save_snapshot(tempfile(fileext = ".png"))
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print an \code{IdfGeometry} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The \code{IdfGeometry} itself, invisibly.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
geom$print()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfGeometry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
