% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idfobjects.R
\docType{class}
\name{IdfObjects}
\alias{IdfObjects}
\title{A Collection of EnergyPlus Objects}
\description{
\code{IdfObjects} is a lazy collection view for multiple \link{IdfObject}s in the same
parent \link{Idf}. It stores object IDs and the parent \code{Idf}, and only creates
individual \link{IdfObject} instances when requested.
}
\seealso{
\link{Idf} class and \link{IdfObject} class
}
\author{
Hongyuan Jia
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IdfObjects-new}{\code{IdfObjects$new()}}
\item \href{#method-IdfObjects-version}{\code{IdfObjects$version()}}
\item \href{#method-IdfObjects-parent}{\code{IdfObjects$parent()}}
\item \href{#method-IdfObjects-length}{\code{IdfObjects$length()}}
\item \href{#method-IdfObjects-id}{\code{IdfObjects$id()}}
\item \href{#method-IdfObjects-object_id}{\code{IdfObjects$object_id()}}
\item \href{#method-IdfObjects-name}{\code{IdfObjects$name()}}
\item \href{#method-IdfObjects-object_name}{\code{IdfObjects$object_name()}}
\item \href{#method-IdfObjects-class_name}{\code{IdfObjects$class_name()}}
\item \href{#method-IdfObjects-group_name}{\code{IdfObjects$group_name()}}
\item \href{#method-IdfObjects-object}{\code{IdfObjects$object()}}
\item \href{#method-IdfObjects-objects}{\code{IdfObjects$objects()}}
\item \href{#method-IdfObjects-slice}{\code{IdfObjects$slice()}}
\item \href{#method-IdfObjects-to_table}{\code{IdfObjects$to_table()}}
\item \href{#method-IdfObjects-to_string}{\code{IdfObjects$to_string()}}
\item \href{#method-IdfObjects-print}{\code{IdfObjects$print()}}
\item \href{#method-IdfObjects-clone}{\code{IdfObjects$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-new"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-new}{}}}
\subsection{Method \code{new()}}{
Create an \code{IdfObjects} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$new(object, class = NULL, parent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{An integer vector of object IDs.}

\item{\code{class}}{An integer vector of class indexes. If \code{NULL}, class
indexes are inferred from \code{object}.}

\item{\code{parent}}{An \link{Idf} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
It is not recommended to directly use \verb{$new()} to create an
\code{IdfObjects} object. Use \link{Idf} methods such as
\href{../../eplusr/html/Idf.html#method-objects}{\code{Idf$objects()}},
\href{../../eplusr/html/Idf.html#method-objects_in_class}{\code{Idf$objects_in_class()}},
and equivalent helpers instead.
}

\subsection{Returns}{
An \code{IdfObjects} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-version"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-version}{}}}
\subsection{Method \code{version()}}{
Get the version of the parent \link{Idf}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$version()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link[base:numeric_version]{base::numeric_version()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-parent"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-parent}{}}}
\subsection{Method \code{parent()}}{
Get the parent \link{Idf}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$parent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An \link{Idf} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-length"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-length}{}}}
\subsection{Method \code{length()}}{
Get the number of objects in the collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$length()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A single integer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-id"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-id}{}}}
\subsection{Method \code{id()}}{
Get object IDs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-object_id"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-object_id}{}}}
\subsection{Method \code{object_id()}}{
Get object IDs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$object_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-name"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-name}{}}}
\subsection{Method \code{name()}}{
Get object names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-object_name"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-object_name}{}}}
\subsection{Method \code{object_name()}}{
Get object names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$object_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-class_name"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-class_name}{}}}
\subsection{Method \code{class_name()}}{
Get class names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$class_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-group_name"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-group_name}{}}}
\subsection{Method \code{group_name()}}{
Get group names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$group_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-object"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-object}{}}}
\subsection{Method \code{object()}}{
Extract one \link{IdfObject} from the collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$object(which)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which}}{A single integer position or object name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \link{IdfObject} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-objects"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-objects}{}}}
\subsection{Method \code{objects()}}{
Materialize the collection as a list of \link{IdfObject} objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$objects(which = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which}}{Optional integer positions, logical indexes, or object
names used to subset before materializing. Default: \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of \link{IdfObject} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-slice"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-slice}{}}}
\subsection{Method \code{slice()}}{
Extract a subset as another \code{IdfObjects} collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$slice(which)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which}}{Integer positions, logical indexes, or object names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{IdfObjects} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-to_table"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-to_table}{}}}
\subsection{Method \code{to_table()}}{
Format the collection as a data.frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$to_table(
  string_value = TRUE,
  unit = TRUE,
  wide = FALSE,
  align = FALSE,
  all = FALSE,
  group_ext = c("none", "group", "index"),
  force = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{string_value}}{If \code{TRUE}, all field values are returned as
character. If \code{FALSE}, \code{value} column in returned
\link[data.table:data.table]{data.table} is a list column with
each value stored as corresponding type. Note that if the
value of numeric field is set to \code{"Autosize"} or
\code{"Autocalculate"}, it is left as it is, leaving the returned
type being a string instead of a number. Default: \code{TRUE}.}

\item{\code{unit}}{Only applicable when \code{string_value} is \code{FALSE}. If
\code{TRUE}, values of numeric fields are assigned with units using
\code{\link[units:units]{units::set_units()}} if applicable. This requires the suggested
package \code{units}. Default: \code{TRUE}.}

\item{\code{wide}}{Only applicable if target objects belong to a same class.
If \code{TRUE}, a wide table will be returned, i.e. first three
columns are always \code{id}, \code{name} and \code{class}, and then every
field in a separate column. Default: \code{FALSE}.}

\item{\code{align}}{If \code{TRUE}, all objects in the same class will have the
same field number. Default: \code{FALSE}.}

\item{\code{all}}{If \code{TRUE}, all available fields defined in IDD for the
class that objects belong to will be returned. Default:
\code{FALSE}.}

\item{\code{group_ext}}{Should be one of \code{"none"}, \code{"group"} or \code{"index"}.
If not \code{"none"}, \code{value} column in returned
\code{\link[data.table:data.table]{data.table::data.table()}} will be converted into a list.
If \code{"group"}, values from extensible fields will be grouped by the
extensible group they belong to. If \code{"index"}, values from
extensible fields will be grouped by the extensible field
indice they belong to. Default: \code{"none"}.}

\item{\code{force}}{If \code{TRUE}, the object IDs are not checked against the
parent \code{Idf} before table generation. Default: \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link[data.table:data.table]{data.table}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-to_string"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-to_string}{}}}
\subsection{Method \code{to_string()}}{
Format the collection as IDF strings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$to_string(
  comment = TRUE,
  header = FALSE,
  format = eplusr_option("save_format"),
  leading = 4L,
  sep_at = 29L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{If \code{FALSE}, all comments will not be included.
Default: \code{TRUE}.}

\item{\code{header}}{If \code{TRUE}, the IDF header will be included. Default:
\code{FALSE}.}

\item{\code{format}}{Specific format used when formatting. For details,
see \href{../../eplusr/html/Idf.html#method-save}{\code{Idf$save()}}.}

\item{\code{leading}}{Leading spaces added to each field. Default: \code{4L}.}

\item{\code{sep_at}}{The character width to separate value string and field
string. Default: \code{29L}, the same as IDF Editor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-print"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-print}{}}}
\subsection{Method \code{print()}}{
Print the collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$print(zoom = c("object", "field"), order = TRUE, n = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zoom}}{Control how detailed the collection should be printed.
Should be one of \code{"object"} and \code{"field"}. Default:
\code{"object"}.}

\item{\code{order}}{If \code{TRUE}, objects are printed in collection order. If
\code{FALSE}, objects are grouped and ordered by class. Default:
\code{TRUE}.}

\item{\code{n}}{Maximum number of objects to print. \code{NULL} prints all
objects. Default: \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{IdfObjects} object itself, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdfObjects-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IdfObjects-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdfObjects$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
