% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplus_read.R
\name{import_epg}
\alias{import_epg}
\title{Import EnergyPlus .epg group file}
\usage{
import_epg(epg)
}
\arguments{
\item{epg}{A file path of EnergyPlus .epg file.}
}
\value{
A tibble.
}
\description{
\code{import_epg} returns a tibble which contains the necessary infomation in
order to run simulations using \code{\link{run_epg}}.
}
