% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\alias{install_eplus}
\title{Install EnergyPlus}
\usage{
install_eplus(ver = "latest", force = FALSE)
}
\arguments{
\item{ver}{The EnergyPlus version number, e.g., \code{8.7}; the special value
\code{latest} means the latest version (fetched from Github
releases).}

\item{force}{Whether to install EnergyPlus even if it has already been
installed.}
}
\description{
Download specified version of EnergyPlus for your platform from Github and
install it.
}
\details{
This function will try to install EnergyPlus into the default location, e.g.
\file{C:/EnergyPlusVX-X-0} on Windows, \file{/usr/local/EnergyPlus-X-Y-0} on
Linux, and \file{/Applications/EnergyPlus-X-Y-0} on MacOS.

Note that the installation process requires administrative privileges
required during the installation and you may need to pass password to enable
sudo commands used for installation.
}
