% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{level_checks}
\alias{level_checks}
\title{Show components of validation strictness level}
\usage{
level_checks(level = eplusr_option("validate_level"))
}
\arguments{
\item{level}{Should be one of \code{"none"}, \code{"draft"}, \code{"final"} or an output of
\code{\link[=custom_validate]{custom_validate()}}.}
}
\value{
A named list with 10 elements, e.g. \code{required_object},
\code{unique_object}, \code{unique_name}, \code{extensible}, \code{required_field}, \code{auto_field},
\code{type}, \code{choice}, \code{range} and \code{reference}. For the meaning of each validation
component, see\code{\link[=custom_validate]{custom_validate()}}.
}
\description{
\code{level_checks()} takes input of a built in validation level or a custom
validation level and returns a list with all validation components that level
contains.
}
\examples{
level_checks("draft")
level_checks("final")
level_checks(custom_validate(auto_field = TRUE))
level_checks(eplusr_option("validate_level"))
}
