% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplus_data.R
\name{long_table}
\alias{long_table}
\title{Format EnergyPlus output and meter result into a long table.}
\usage{
long_table(data)
}
\arguments{
\item{data}{A data.table object containing EnergyPlus standard output or
meter result.}
}
\value{
A data.table with long table format.
}
\description{
\code{long_table} takes an EnergyPlus "output" or "meter" result as an input
and returns a long table with first column being the POSIXt column, and next
'component' indicating energy consumption components, 'type' indicating
energy types (e.g.  Electricity, and Gas), 'value' indicating the value of
the component, 'unit' indicating the unit of energy used, and 'timestep'
indicating the tiem step of data collected. A meter output from a
10-min-timestep annual simulation will takes about 5 seconds to load.  So,
use with caution.
}
