% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{na_replace}
\alias{na_replace}
\title{Replace NA, NAN or Inf in a data.table.}
\usage{
na_replace(data, col_pattern, type = "na", replacement = 0L)
}
\arguments{
\item{data}{A data.frame or data.table object to manipulate. If a data.frame,
it will be converted into a data.table.}

\item{col_pattern}{A RegEx used for column subsetting.}

\item{type}{A character indicates the type of nonsense data. Currently,
\code{NA}, \code{NAN}, and \code{Inf} are support.}

\item{replacement}{A single character or number used to replace the nonsense
data.}
}
\value{
A data.table with nonsense data replaced.
}
\description{
Replace NA, NAN or Inf in a data.table.
}
