% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nmbe}
\alias{nmbe}
\title{Calculate NMBE between two numeric vector.}
\usage{
nmbe(x, y, na.rm = TRUE, ...)
}
\arguments{
\item{x, y}{A numeric vector to calculate with.}

\item{na_rm}{If TRUE, NA values will be removed.}
}
\value{
The NMBE of x to y.
}
\description{
Calculate NMBE between two numeric vector.
}
