% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wthr_download.R
\docType{data}
\name{page_main}
\alias{page_main}
\title{Download weather file from EnergyPlus website}
\format{An object of class \code{character} of length 1.}
\usage{
page_main
}
\arguments{
\item{search}{A string used to search in the weather database.}

\item{savedir}{A directory where the weather file will be downloaded to.}

\item{type}{The desired file format. Should be one of "all", "ddy", "epw", or
"stat".}
}
\description{
Weather data for more than 2100 locations are now available in EnergyPlus
weather format — 1042 locations in the USA, 71 locations in Canada, and more
than 1000 locations in 100 other countries throughout the world. The weather
data are arranged by World Meteorological Organization region and Country.
\code{wthr_download} will search the weather database in EnergyPlus website,
return the results and automatically download the weather file.
}
\examples{
wthr_download("chongqing")
}
\keyword{datasets}
