% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{param_job}
\alias{param_job}
\title{Create An EnergyPlus Parametric Simulation Job}
\usage{
param_job(idf, epw)
}
\arguments{
\item{idf}{A path to EnergyPlus IDF or IMF file or an \code{Idf} object.}

\item{epw}{A path to EnergyPlus EPW file or an \code{Epw} object.}
}
\value{
A \code{ParametricJob} object.
}
\description{
\code{param_job()} takes an IDF and EPW as input and returns a \code{ParametricJob}.
For details on \code{ParametricJob}, please see \link[=param]{ParametricJob class}.
}
\seealso{
\code{\link[=eplus_job]{eplus_job()}} for creating an EnergyPlus single simulation job.
}
