% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.R
\name{parse_idf}
\alias{parse_idf}
\title{Parse EnergyPlus models}
\usage{
parse_idf(idf_str, idd)
}
\arguments{
\item{idf_str}{A character vector created using \code{read_idf}.}

\item{idd}{An IDD object created using \code{parse_idd}.}
}
\value{
A list contains the IDF version, option data, class data, value data,
comment data and field reference data.
}
\description{
Parse EnergyPlus models
}
