% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err.R
\name{print.ErrFile}
\alias{print.ErrFile}
\title{Print EnergyPlus Error File}
\usage{
\method{print}{ErrFile}(x, brief = FALSE, info = TRUE, ...)
}
\arguments{
\item{x}{An \code{ErrFile} created using \code{\link[=read_err]{read_err()}}.}

\item{brief}{If \code{TRUE}, only summary data is printed. Default: \code{FALSE}.}

\item{info}{If \code{FALSE}, informative messages are excluded. Only warnings and
errors are printed. Default: \code{TRUE}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An \code{ErrFile} object, invisibly.
}
\description{
\code{ErrFile} is mainly used to extract and print data in an EnergyPlus Error
File (\code{.err}).
}
