% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err.R
\name{print.TransitionErrFile}
\alias{print.TransitionErrFile}
\title{Print EnergyPlus Transition Error File}
\usage{
\method{print}{TransitionErrFile}(x, brief = FALSE, info = TRUE, ...)
}
\arguments{
\item{x}{An \code{TransitionErrFile} created using \code{\link[=read_err]{read_err()}}.}

\item{brief}{If \code{TRUE}, only summary data is printed. Default: \code{FALSE}.}

\item{info}{If \code{FALSE}, informative messages are excluded. Only warnings and
errors are printed. Default: \code{TRUE}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An \code{TransitionErrFile} object, invisibly.
}
\description{
\code{TransitionErrFile} is mainly used to extract and print data in an EnergyPlus
Transition Error File (\code{.vcperr}).
}
