% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl-idf.R
\name{purge_idf_object}
\alias{purge_idf_object}
\title{Purge not-used resource objects}
\usage{
purge_idf_object(idd_env, idf_env, dt_object)
}
\arguments{
\item{idd_env}{An environment or list contains IDD tables including class,
field, and reference.}

\item{idf_env}{An environment or list contains IDF tables including object,
value, and reference.}

\item{dt_object}{A \code{\link[data.table:data.table]{data.table::data.table()}} that contains object data.
Usually created using \code{\link[=expand_idf_dots_value]{expand_idf_dots_value()}}.}

\item{level}{Validate level. Default: \code{eplusr_option("validate_level")}.}
}
\value{
The modified whole IDF data in a named list of 3
\code{\link[data.table:data.table]{data.table::data.table()}}s, i.e. \code{object}, \code{value} and \code{reference}.
}
\description{
Purge not-used resource objects
}
\keyword{internal}
