% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epw.R
\name{read_epw}
\alias{read_epw}
\title{Read and Parse EnergyPlus Weather File (EPW)}
\usage{
read_epw(path)
}
\arguments{
\item{path}{A path of an EnergyPlus \code{EPW} file.}
}
\value{
An \code{Epw} object.
}
\description{
\code{read_epw()} parses an EPW file and returns an \code{Epw} object. The parsing
process is extreme inspired by \href{https://openstudio-sdk-documentation.s3.amazonaws.com/cpp/OpenStudio-2.5.1-doc/utilities/html/classopenstudio_1_1_epw_file.html}{OpenStudio utilities library}
with some simplifications. For more details on \code{Epw}, please see \link[=epw]{Epw Class}.
}
\seealso{
\link[=epw]{Epw class}
}
