% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.R
\name{read_idf}
\alias{read_idf}
\title{Read an EnergyPlus Input Data File (IDF)}
\usage{
read_idf(path, idd = NULL)
}
\arguments{
\item{path}{A path to an EnergyPlus IDF file or a string that can be parsed as
an IDF. The file extension does not matter. So models stored in \code{TXT}
format are still able to correctly be parsed.}

\item{idd}{A path to an \code{Energy+.idd} file. If NULL, the pre-parsed
\code{Energy+.idd} files stored internally from EnergyPlus v8.3 to 8.8 will be
used according to the version of input model, if possible.}
}
\value{
An \code{Idf} object.
}
\description{
\code{read_idf} takes an EnergyPlus Input Data File (IDF) as input and returns an
\code{Idf} object. For more details on \code{Idf} object, please see \link{idf}.
}
