% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_jeplus}
\alias{read_jeplus}
\title{Import jEPlus .json type project.}
\usage{
read_jeplus(json)
}
\arguments{
\item{json}{A file path of a .json file.}
}
\value{
A list containing project info.
}
\description{
\code{read_jeplus} takes a file path of an .json type project of jEPlus,
and return a list containing model paths, weather paths, parametric fields,
and parametric values. The returned list will have an attribute 'job_type'
with value 'jeplus' which will be used when running jobs using
\link{\code{run_job}}.
}
