% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplus_read.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read_csv}
\title{Extract EnergyPlus output info}
\arguments{
\item{file}{Either a folder/file path which contains EnergyPlus simulation
results, a EPAT job_index.csv}
}
\description{
\code{read_outinfo} returns a tibble contains the key information of
EnergyPlus \code{Output:Variable} and \code{Output:Meter}.
}
\examples{

}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{readr}{\code{\link[readr]{read_csv}}}
}}

