% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl-idf.R
\name{remove_duplicated_objects}
\alias{remove_duplicated_objects}
\title{Remove duplicated objects in inputs}
\usage{
remove_duplicated_objects(idd_env, idf_env, dt_object, dt_value)
}
\arguments{
\item{idd_env}{An environment or list contains IDD tables including class,
field, and reference.}

\item{idf_env}{An environment or list contains IDF tables including object,
value, and reference.}

\item{dt_object}{A \code{\link[data.table:data.table]{data.table::data.table()}} that contains object data.}

\item{dt_value}{A \code{\link[data.table:data.table]{data.table::data.table()}} that contains value data.}
}
\value{
The modified input data in a named list of 2
\code{\link[data.table:data.table]{data.table::data.table()}}s, i.e. \code{object} and \code{value}.
}
\description{
Remove duplicated objects in inputs
}
\keyword{internal}
