% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl-idf.R
\name{remove_empty_fields}
\alias{remove_empty_fields}
\title{Remove trailing empty object fields}
\usage{
remove_empty_fields(idd_env, idf_env, dt_value)
}
\arguments{
\item{idd_env}{An environment or list contains IDD tables including class,
field, and reference.}

\item{idf_env}{An environment or list contains IDF tables including object,
value, and reference.}

\item{dt_value}{A \code{\link[data.table:data.table]{data.table::data.table()}} that contains value data.}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}}
}
\description{
Remove trailing empty object fields
}
\keyword{internal}
