% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl-idf.R
\name{rename_idf_object}
\alias{rename_idf_object}
\title{Rename existing objects}
\usage{
rename_idf_object(
  idd_env,
  idf_env,
  dt_object,
  level = eplusr_option("validate_level")
)
}
\arguments{
\item{idd_env}{An environment or list contains IDD tables including class,
field, and reference.}

\item{idf_env}{An environment or list contains IDF tables including object,
value, and reference.}

\item{dt_object}{A \code{\link[data.table:data.table]{data.table::data.table()}} that contains object data.
Usually created using \code{\link[=expand_idf_dots_name]{expand_idf_dots_name()}}.}

\item{level}{Validate level. Default: \code{eplusr_option("validate_level")}.}
}
\value{
The modified object data in a named list of 3
\code{\link[data.table:data.table]{data.table::data.table()}}s, i.e. \code{object}, \code{value}, \code{reference}.
}
\description{
Rename existing objects
}
\note{
\itemize{
\item The \code{reference} table in the returned list only contains the reference-by
map, indicating which values have been updated in the main \code{value} table in
\code{idf_env}.
\item The \code{value} table in input \code{idf_env} could be modified if input objects are
referenced by other objects. The \code{reference} table in the returned list
tells the id actual values modified
}
}
\keyword{internal}
