% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{resample}
\alias{resample}
\title{Resample data according to given time step.}
\usage{
resample(data, base = NULL, new = NULL, step = "month", toward = c("up",
  "down", "center"), drop = FALSE, fun = mean, ...)
}
\arguments{
\item{data}{A data frame containing at least one datetime column.}

\item{base}{A character indicates the name of datetime column.  If not given,
it will be auto-detected.}

\item{new}{The column name of the added variable. If 'NULL' it will be
the name of the original datetime variable with the interval name added to
it, separeted by an underscore.}

\item{step}{The timestep of the added datetime variable, which should be
higher than the interval of the input datetime variable. If 'NULL', it will
be one level higher than the interval of the input datetime variable. It can
also be an integer indicates the interval of minutes. In conclusion,
'interval' should be a number or one of c("year", "quarter", "month", "week",
"yday", "mday", "wday", "day", "date", "hour", "min", "second").}

\item{drop}{If TRUE, all non-numeric columns will be dropped during
resampling. Otherwise, they will be used as group columns.}
}
\description{
\code{resample} adds a padding time column and aggregate time series data
according to given time step.
}
