% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{row_select}
\alias{row_select}
\title{Row subsetting using RegEx in a data.table.}
\usage{
row_select(data, based_col = "Variable", row_pattern = "")
}
\arguments{
\item{data}{A data.frame or data.table to subset culumns from.}

\item{based_col}{A character column name used for row subsetting.}

\item{row_pattern}{A RegEx used for row subsetting.}
}
\value{
It returns a data.table with subsetted rows.
}
\description{
\code{row_select} is a function for easy row subsetting using RegEx in one
column in a data.table, especially when the target column is a character
type. It will first check whether the input is a data.table object or not. If
not, it will convert it into a data.table and all row names will be lost.
}
\examples{
row_select(iris, "Species", "^(?!setosa)")
}
