% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl-idf.R
\name{set_idf_object}
\alias{set_idf_object}
\title{Modifying existing objects}
\usage{
set_idf_object(
  idd_env,
  idf_env,
  dt_object,
  dt_value,
  empty = FALSE,
  level = eplusr_option("validate_level")
)
}
\arguments{
\item{idd_env}{An environment or list contains IDD tables including class,
field, and reference.}

\item{idf_env}{An environment or list contains IDF tables including object,
value, and reference.}

\item{dt_object}{A \code{\link[data.table:data.table]{data.table::data.table()}} that contains object data.
Usually created using \code{\link[=expand_idf_dots_value]{expand_idf_dots_value()}}.}

\item{dt_value}{A \code{\link[data.table:data.table]{data.table::data.table()}} that contains value data.
Usually created using \code{\link[=expand_idf_dots_value]{expand_idf_dots_value()}}.}

\item{empty}{If \code{FALSE}, not required empty fields will be removed.
Default: \code{FALSE}.}

\item{level}{Validate level. Default: \code{eplusr_option("validate_level")}.}
}
\value{
The modified object data in a named list of 3
\code{\link[data.table:data.table]{data.table::data.table()}}s, i.e. \code{object}, \code{value} and \code{reference}.
}
\description{
Modifying existing objects
}
\keyword{internal}
