% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplus_app.R
\name{show_output}
\alias{show_output}
\title{A Shiny Gadget for interactively showing EnergyPlus output.}
\usage{
show_output(data, state = NULL, group = NULL, out = c("ggplot2", "plotly",
  "dygraphs", "state", "long_table", "wide_table", "output_info"))
}
\arguments{
\item{data}{A data.frame containing EnergyPlus output.}

\item{state}{A list containing all or some input values of the Shiny Gadget.
This simulates some functionality of Shiny Bookmark.}

\item{group}{A single character string indicating the group column name if
the input data contains multiple EnergyPlus simulation results.}

\item{out}{The return of the Shiny Gadget.}
}
\value{
Depands on the \code{out}.
}
\description{
\code{show_output} takes an EnergyPlus output, e.g. variables, meters, and
launch a shiny gadget for interactively subsetting or plotting the data.
}
