% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{site_to_src}
\alias{site_to_src}
\title{Tramsform EnergyPlus meter output from site energy to source energy.}
\usage{
site_to_src(data, ele_pattern = "electricity", gas_pattern = "gas",
  ele_fct = 3.095, gas_fct = 1.092, digits = 4, to_GJ = FALSE)
}
\arguments{
\item{data}{A data.table object containing EnergyPlus standard meter result.}

\item{ele_pattern}{A RegEx to extract columns of electricity consumption.}

\item{gas_pattern}{A RegEx to extract columns of gas consumption.}

\item{ele_fct}{The site to source convert factor of electricity. Default is 3.095
which is the standard value in California, USA.}

\item{gas_fct}{The site to source convert factor of gas. Default is 1.092
which is the standard value in California, USA.}

\item{digits}{A integer indicates the digits of calculated results.}

\item{to_GJ}{If TRUE, the unit of energy consumption will be converted from
Joule to Gigajoule.}
}
\value{
A data.table containing EnergyPlus meter results with site energy
converted into source energy.
}
\description{
Tramsform EnergyPlus meter output from site energy to source energy.
}
