% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl-idf.R
\name{standardize_idf_value}
\alias{standardize_idf_value}
\title{Standardize Value Data}
\usage{
standardize_idf_value(
  idd_env,
  idf_env,
  dt_value,
  type = c("choice", "reference")
)
}
\arguments{
\item{idd_env}{An environment or list contains IDD tables including class,
field, and reference.}

\item{idf_env}{An environment or list contains IDF tables including object,
value, and reference.}

\item{type}{A character vector to specify what type of values to be
standardized. Should be a subset of \code{c("choice", "reference")}.
Default: \code{c("choice", "reference")}.}

\item{class}{An integer vector of valid class indexes or a character vector
of valid class names. Default: \code{NULL}.}

\item{object}{An integer vector of valid object IDs or a character vector
of valid object names. Default: \code{NULL}.}

\item{field}{An integer vector of valid field indexes or a character
vector of valid field names (can be in in underscore style). \code{class}
and \code{field} should have the same length.}
}
\value{
A data.table
}
\description{
Standardize Value Data
}
\keyword{internal}
