% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tg_to_tr}
\alias{tg_to_tr}
\title{Calculate Mean Radiant Temperature (MRT).}
\usage{
tg_to_tr(tg, ta, d = 0.038, e = 0.9, v = 0.01)
}
\arguments{
\item{tg}{Global Temperature}

\item{ta}{Mean air temperature measured at the same height of \code{tg}.}

\item{d}{The dimeter of the globe.}

\item{e}{The emissivity of the globe.}

\item{v}{The air velocity.}
}
\value{
Calculated mean radiant temperature.
}
\description{
\code{tg_to_tr} returns
}
