% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{time_from_eplus}
\alias{time_from_eplus}
\title{Transform EnergyPlus DateTime format into POSIXct class.}
\usage{
time_from_eplus(x, year = current_year(), tz = Sys.timezone())
}
\arguments{
\item{x}{An EnergyPlus DateTime format string vector.}

\item{year}{Year number.}

\item{tz}{Time zone of the result POSIXct sequence.}
}
\value{
A POSIXct sequence.
}
\description{
\code{time_from_eplus} takes an EnergyPlus DateTime format string, e.g.
" %m/%d  %H:%M:%d" into a POSIXct sequence in R.
}
