% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{time_to_eplus}
\alias{time_to_eplus}
\title{Transform a POSIXct object to EnergyPlus DateTime format string.}
\usage{
time_to_eplus(x)
}
\arguments{
\item{x}{A POSIXct object.}
}
\value{
A string vector
}
\description{
\code{time_to_eplus} takes a POSIXct and formats it into a string with
EnergyPlus DateTime format, e.g." %m/%d  %H:%M:%d".
}
