% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{transition}
\alias{transition}
\title{Perform version transition of EnergyPlus model}
\usage{
transition(idf, ver, keep_all = FALSE)
}
\arguments{
\item{idf}{An \link{Idf} object or a path of IDF file.}

\item{ver}{A valid EnergyPlus version, e.g. \code{9}, \code{8.8}, or \code{"8.8.0"}.}

\item{keep_all}{If \code{TRUE}, a list will be return which contains all
\link{Idf} objects of intermediate versions. The list will be named using first
two number of that version, e.g. \code{8.1}, \code{8.2}.}
}
\value{
An \link{Idf} object if \code{keep_all} is \code{FALSE} or a list of \link{Idf} objects
if \code{keep_all} is \code{TRUE}.
}
\description{
\code{transition()} takes an \link{Idf} object or a path of IDF file and a target
version, performs version transitions and returns an \link{Idf} object of
specified version.
}
