% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{with_speed}
\alias{with_speed}
\title{Evaluates an expression without checking}
\usage{
with_speed(expr)
}
\arguments{
\item{expr}{An expression to be evaluated.}
}
\description{
Evaluates an expression without checking
}
\keyword{internal}
