% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{yhour}
\alias{yhour}
\title{Get the hour number in a year.}
\usage{
yhour(x, one_year = FALSE, no_leap = FALSE)
}
\arguments{
\item{x}{A datetime object.}

\item{one_year}{If TRUE, the calculation will based on the earliest year in
the input}

\item{no_leap}{If TRUE, and the year is a leap year, Feb 29th will be
neglected during the calculation.}
}
\value{
A numeric vector indicates the hour number in the year.
}
\description{
\code{yhour} returns the hour number in a year.
}
