% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{add_time}
\alias{add_time}
\title{Add a padding time column for further aggregation.}
\usage{
add_time(data, base = NULL, new = NULL, step, toward = c("up", "down",
  "center"), no_leap = FALSE)
}
\arguments{
\item{data}{A data frame containing at least one datetime column.}

\item{base}{A character indicates the name of datetime column.  If not given,
it will be auto-detected.}

\item{new}{The column name of the added variable. If 'NULL' it will be
the name of the original datetime variable with the interval name added to
it, separeted by an underscore.}

\item{step}{The timestep of the added datetime variable, which should be
higher than the interval of the input datetime variable. If 'NULL', it will
be one level higher than the interval of the input datetime variable. It can
also be an integer indicates the interval of minutes. In conclusion,
'interval' should be a number or one of c("year", "quarter", "month", "week",
"yday", "mday", "wday", "day", "date", "hour", "min", "second").}

\item{one_year}{If TRUE, the (year+1)-01-01 will be replaced with
(year)-12-31.}
}
\description{
\code{add_time} adds a padding time column in a data.frame for further easy
data aggregation. It can also replace the datetime that is (year+1)-01-01
with (year)-12-31.
}
