% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{clean_wd}
\alias{clean_wd}
\title{Clean working directory of a previous EnergyPlus simulation}
\usage{
clean_wd(path)
}
\arguments{
\item{path}{An \code{.idf} or \code{.imf} file path.}
}
\description{
Clean working directory of an EnergyPlus simulation by deleting all input and
output files of previous simulation.
}
