% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_wd}
\alias{clean_wd}
\title{Clean an working directory of EnergyPlus}
\usage{
clean_wd(path, suffix_type = c("C", "L", "D"), extra = NULL,
  keep_input = FALSE, backup = FALSE, backup_folder = NULL,
  mark = "datetime", newname_mark = NULL)
}
\arguments{
\item{path}{A EnergyPlus working directory.}

\item{backup}{Should be NULL or one of c('input', 'basic', 'all').  If NULL,
none file will be backed up.  If 'input', *.imf, *.idf, *.epw, *.epmidf will
be backed up.  If 'basic', *.imf, *.idf, *.ewp, *.epmidf, and results of
summary report, meter output and variable output will be backed up.  If
'all', all EnergyPlus input and output files will be backed up.}

\item{backup_folder}{A character indicates the name prefix of the folder that
will be auto-generated in current working directory to store backed-up files.
The full folder name will be ('backup_folder'_(suffix)). The suffix is a
character that shows the current date and time in format like
'Apr06_01h_02m'.  If this this folder already, which means that you run this
function multiple times within 1 minute, the folder name will be the decribed
above with 6 random characters.}

\item{rename}{If TRUE, the backed-up files will be renamed with the same
suffix as backup_folder.}
}
\description{
\code{clean_wd} is a function to clean the working directory of EnergyPlus
after backing up specifed files. All other files that are not parts of
EnergyPlus input or output files will be left as it is.  This is also true if
you rename those files of EnergyPlus before you run this function. Currently,
files work as input schedule files are not supported, and will be left
untouched.
}
