% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{col_select}
\alias{col_select}
\title{Column selection using RegEx in a data.table.}
\usage{
col_select(data, col = NULL, by = NULL, regex = TRUE,
  ignore_case = TRUE, invert = FALSE)
}
\arguments{
\item{data}{A data.frame or data.table to subset culumns from.}

\item{col}{A RegEx used for column subsetting. If blank, all columns be
returned.}

\item{by}{A character vector used for grouped column subsetting using the
feature of \code{by} argument in data.table. Especially useful when you want
grouping or cases when the \code{by} columns are not easy to be combined with
\code{col}. NOTE: \code{by} is not a regex, but a character vector.}

\item{regex}{If FALSE, \code{col} is treated as a character vector of
column names, not a regex.}

\item{ignore_case}{A logical value. If FALSE, the pattern matching is case
sensitive and if TRUE, case is ignored during matching. Default is TRUE.
\code{ignore_case} only takes effect on \code{col} not \code{by}.}

\item{invert}{A logical value. If TRUE return column names that do not match.
Default is FALSE. \code{invert} only takes effect on \code{col} not
\code{by}.}
}
\value{
A data.table that only contains the subsetted columns.
}
\description{
\code{col_select} will first check whether the input is a data.table object
or not. If not, it will convert it into a data.table and all row names will
be lost.
}
\examples{
col_select(mtcars, "t$")
}
