% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl-idf.R
\name{duplicated_idf_object}
\alias{duplicated_idf_object}
\title{Determine duplicate objects}
\usage{
duplicated_idf_object(idd_env, idf_env, dt_object)
}
\arguments{
\item{idd_env}{An environment or list contains IDD tables including class,
field, and reference.}

\item{idf_env}{An environment or list contains IDF tables including object,
value, and reference.}

\item{dt_object}{A \code{\link[data.table:data.table]{data.table::data.table()}} that contains object data.}
}
\value{
A same \code{\link[data.table:data.table]{data.table::data.table()}} as input \code{dt_object} (updated by
reference) with appended integer column \code{unique_object_id} indicating the
object is a duplicated one of that object.
}
\description{
Determine duplicate objects
}
\keyword{internal}
