% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{energyplus}
\alias{energyplus}
\alias{EPMacro}
\alias{ExpandObjects}
\alias{Basement}
\alias{Slab}
\alias{EnergyPlus}
\alias{convertESOMTR}
\alias{ReadVarsESO}
\alias{HVAC_Diagram}
\title{Run EnergyPlus}
\usage{
EPMacro(
  model,
  output_dir = NULL,
  output_prefix = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL
)

ExpandObjects(
  model,
  output_dir = NULL,
  output_prefix = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL,
  idd = NULL
)

Basement(
  model,
  weather,
  output_dir = NULL,
  output_prefix = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL,
  idd = NULL
)

Slab(
  model,
  weather,
  output_dir = NULL,
  output_prefix = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL,
  idd = NULL
)

EnergyPlus(
  model,
  weather,
  output_dir = NULL,
  output_prefix = NULL,
  output_suffix = c("C", "L", "D"),
  wait = TRUE,
  echo = TRUE,
  annual = FALSE,
  design_day = FALSE,
  idd = NULL,
  eplus = NULL
)

convertESOMTR(
  eso,
  output_dir = NULL,
  output_prefix = NULL,
  rules = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL
)

ReadVarsESO(
  eso,
  output_dir = NULL,
  output_prefix = NULL,
  output_suffix = c("C", "L", "D"),
  max_col = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL
)

HVAC_Diagram(
  bnd,
  output_dir = NULL,
  output_prefix = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL
)

energyplus(
  model,
  weather,
  output_dir = NULL,
  output_prefix = NULL,
  output_suffix = c("C", "L", "D"),
  epmacro = TRUE,
  expand_obj = TRUE,
  annual = FALSE,
  design_day = FALSE,
  eso_to_ip = FALSE,
  readvars = TRUE,
  wait = TRUE,
  echo = TRUE,
  idd = NULL,
  eplus = NULL
)
}
\description{
Run EnergyPlus
}
\keyword{internal}
