% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{eplus_job}
\alias{eplus_job}
\title{Create an EnergyPlus Simulation Job}
\usage{
eplus_job(idf, epw)
}
\arguments{
\item{idf}{A path to an EnergyPlus IDF or IMF file or an \code{Idf} object.}

\item{epw}{A path to an EnergyPlus EPW file or an \code{Epw} object.}
}
\value{
An \code{EplusJob} object.
}
\description{
\code{eplus_job} takes an IDF and EPW as input, and returns an \code{EplusJob} object
for running EnergyPlus simulation and collecting outputs. For more details,
please see \link{job}.
}
\seealso{
\code{\link[=param_job]{param_job()}} for creating an EnergyPlus parametric job.
}
