% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\docType{class}
\name{EplusJob}
\alias{EplusJob}
\alias{eplus_job}
\title{Run EnergyPlus Simulation and Collect Outputs}
\usage{
eplus_job(idf, epw)
}
\arguments{
\item{idf}{A path to an local EnergyPlus IDF file or an \code{Idf} object.}

\item{epw}{A path to an local EnergyPlus EPW file or an \code{Epw} object. \code{epw}
can also be \code{NULL} which will force design-day-only simulation when
\code{\link[=EplusJob]{$run()}} method is called. Note this needs at least one
\code{Sizing:DesignDay} object exists in the \link{Idf}.}
}
\value{
An \code{EplusJob} object.
}
\description{
\code{EplusJob} class wraps the EnergyPlus command line interface and provides
methods to extract simulation outputs.

\code{eplus_job()} takes an IDF and EPW as input, and returns an \code{EplusJob} object
for running EnergyPlus simulation and collecting outputs.
}
\details{
eplusr uses the EnergyPlus SQL output and the CSV output for extracting
simulation outputs. It will use the CSV output if possible, and will fall
back to the SQL output when CSV output does not exist. Underneath,
\code{\link[data.table:fread]{data.table::fread()}} is used to speed up the process when using CSV output.

\code{EplusJob} has provide some wrappers that do SQL query to get report data
results, i.e. results from \code{Output:Variable} and \verb{Output:Meter*}. But for
\code{Output:Table} results, you have to be familiar with the structure of the
EnergyPlus SQL results, especially for table \emph{"TabularDataWithStrings"}. For
details, please see \emph{"2.20 eplusout.sql"}, especially \emph{"2.20.4.4 TabularData
Table"} in EnergyPlus \emph{"Output Details and Examples"} documentation. An
object in \code{Output:SQLite} with \verb{Option Type} value of \code{SimpleAndTabular} will
be automatically created if it does not exists, to ensure that the output
collection functionality works successfully.

In order to make sure \code{.rdd} (Report Data Dictionary) and \code{.mdd} (Meter Data
Dictionary) files are created during simulation, an object in
\code{Output:VariableDictionary} class with \verb{Key Field} value being \code{IDF} will be
automatically created if it does not exists.

In order to make sure all output variables and meters are placed in one CSV
output, all objects in class \code{Output:Meter:MeterFileOnly}and
\code{Output:Meter:Cumulative:MeterFileOnly} will be converted into \code{Output:Meter}
and \code{Output:Meter:Cumulative}, respectively.
}
\examples{

## ------------------------------------------------
## Method `EplusJob$new`
## ------------------------------------------------

\dontrun{
if (is_avail_eplus(8.8)) {
    idf_name <- "1ZoneUncontrolled.idf"
    epw_name <-  "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"

    idf_path <- file.path(eplus_config(8.8)$dir, "ExampleFiles", idf_name)
    epw_path <- file.path(eplus_config(8.8)$dir, "WeatherData", epw_name)

    # create from local files
    job <- eplus_job(idf_path, epw_path)

    # create from an Idf and an Epw object
    job <- eplus_job(read_idf(idf_path), read_epw(epw_path))
}
}

## ------------------------------------------------
## Method `EplusJob$version`
## ------------------------------------------------

\dontrun{
job$version()
}


## ------------------------------------------------
## Method `EplusJob$path`
## ------------------------------------------------

\dontrun{
job$path()
job$path("idf")
job$path("epw")
}


## ------------------------------------------------
## Method `EplusJob$run`
## ------------------------------------------------

\dontrun{
# only run design day
job$run(NULL)

# specify output directory
job$run(dir = tempdir())

# run in the background
job$run(wait = TRUE)
# see job status
job$status()

# force to kill background job before running the new one
job$run(force = TRUE)

# do not show anything in the console
job$run(echo = FALSE)

# copy external files used in the model to simulation output directory
job$run(copy_external = TRUE)
}


## ------------------------------------------------
## Method `EplusJob$kill`
## ------------------------------------------------

\dontrun{
job$kill()
}


## ------------------------------------------------
## Method `EplusJob$status`
## ------------------------------------------------

\dontrun{
job$status()
}


## ------------------------------------------------
## Method `EplusJob$errors`
## ------------------------------------------------

\dontrun{
job$errors()

# show all information
job$errors(info = TRUE)
}


## ------------------------------------------------
## Method `EplusJob$output_dir`
## ------------------------------------------------

\dontrun{
job$output_dir()

# Below will open output directory
# job$output_dir(open = TRUE)
}


## ------------------------------------------------
## Method `EplusJob$locate_output`
## ------------------------------------------------

\dontrun{
# get the file path of the error file
job$locate_output(".err", strict = FALSE)

# can use to detect if certain output file exists
job$locate_output(".expidf", strict = TRUE)
}


## ------------------------------------------------
## Method `EplusJob$read_rdd`
## ------------------------------------------------

\dontrun{
job$read_rdd()
}


## ------------------------------------------------
## Method `EplusJob$read_mdd`
## ------------------------------------------------

\dontrun{
job$read_mdd()
}


## ------------------------------------------------
## Method `EplusJob$list_table`
## ------------------------------------------------

\dontrun{
job$list_table()
}


## ------------------------------------------------
## Method `EplusJob$read_table`
## ------------------------------------------------

\dontrun{
# read a specific table
job$read_table("Zones")
}


## ------------------------------------------------
## Method `EplusJob$report_data_dict`
## ------------------------------------------------

\dontrun{
job$report_data_dict()
}


## ------------------------------------------------
## Method `EplusJob$report_data`
## ------------------------------------------------

\dontrun{
# read all report data
job$report_data()

# specify output variables using report data dictionary
dict <- job$report_data_dict()
job$report_data(dict[units == "C"])

# specify output variables using 'key_value' and 'name'
job$report_data("environment", "site outdoor air drybulb temperature")

# explicitly specify year value and time zone
job$report_data(dict[1], year = 2020, tz = "Etc/GMT+8")

# explicitly specify case name
job$report_data(dict[1], case = "example")

# get all possible columns
job$report_data(dict[1], all = TRUE)

# return in a format that is similar as EnergyPlus CSV output
job$report_data(dict[1], wide = TRUE)

# return in a format that is similar as EnergyPlus CSV output with
# extra columns
job$report_data(dict[1], wide = TRUE, all = TRUE)

# only get data at the working hour on the first Monday
job$report_data(dict[1], hour = 8:18, day_type = "monday", simulation_days = 1:7)

# only get specified run period data
job$read_table("EnvironmentPeriods") # possible environment name
job$report_data(dict[1], environment_name = "San Francisco Intl Ap CA USA TMY3 WMO#=724940")
# can also be done using 'environment_period_index' column
job$report_data(dict[1], all = TRUE)[environment_period_index == 3L]
}


## ------------------------------------------------
## Method `EplusJob$tabular_data`
## ------------------------------------------------

\dontrun{
# read all tabular data
job$tabular_data()

# explicitly specify data you want
str(job$tabular_data(
    report_name = "AnnualBuildingUtilityPerformanceSummary",
    table_name = "Site and Source Energy",
    column_name = "Total Energy",
    row_name = "Total Site Energy"
))

# get tabular data in wide format and coerce numeric values
str(job$tabular_data(
    report_name = "AnnualBuildingUtilityPerformanceSummary",
    table_name = "Site and Source Energy",
    column_name = "Total Energy",
    row_name = "Total Site Energy",
    wide = TRUE, string_value = FALSE
))
}


## ------------------------------------------------
## Method `EplusJob$print`
## ------------------------------------------------

\dontrun{
job$print()
}

}
\seealso{
\link{ParametricJob} class for EnergyPlus parametric simulations.

\code{\link[=param_job]{param_job()}} for creating an EnergyPlus parametric job.
}
\author{
Hongyuan Jia
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EplusJob$new()}}
\item \href{#method-version}{\code{EplusJob$version()}}
\item \href{#method-path}{\code{EplusJob$path()}}
\item \href{#method-run}{\code{EplusJob$run()}}
\item \href{#method-kill}{\code{EplusJob$kill()}}
\item \href{#method-status}{\code{EplusJob$status()}}
\item \href{#method-errors}{\code{EplusJob$errors()}}
\item \href{#method-output_dir}{\code{EplusJob$output_dir()}}
\item \href{#method-locate_output}{\code{EplusJob$locate_output()}}
\item \href{#method-read_rdd}{\code{EplusJob$read_rdd()}}
\item \href{#method-read_mdd}{\code{EplusJob$read_mdd()}}
\item \href{#method-list_table}{\code{EplusJob$list_table()}}
\item \href{#method-read_table}{\code{EplusJob$read_table()}}
\item \href{#method-report_data_dict}{\code{EplusJob$report_data_dict()}}
\item \href{#method-report_data}{\code{EplusJob$report_data()}}
\item \href{#method-tabular_data}{\code{EplusJob$tabular_data()}}
\item \href{#method-print}{\code{EplusJob$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an \code{EplusJob} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$new(idf, epw)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{idf}}{Path to an local EnergyPlus IDF file or an \link{Idf} object.}

\item{\code{epw}}{Path to an local EnergyPlus EPW file or an \link{Epw} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{EplusJob} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
if (is_avail_eplus(8.8)) {
    idf_name <- "1ZoneUncontrolled.idf"
    epw_name <-  "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"

    idf_path <- file.path(eplus_config(8.8)$dir, "ExampleFiles", idf_name)
    epw_path <- file.path(eplus_config(8.8)$dir, "WeatherData", epw_name)

    # create from local files
    job <- eplus_job(idf_path, epw_path)

    # create from an Idf and an Epw object
    job <- eplus_job(read_idf(idf_path), read_epw(epw_path))
}
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-version"></a>}}
\if{latex}{\out{\hypertarget{method-version}{}}}
\subsection{Method \code{version()}}{
Get the version of IDF in current job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$version()}\if{html}{\out{</div>}}
}

\subsection{Details}{
\verb{$version()} returns the version of IDF that current \code{EplusJob} uses.
}

\subsection{Returns}{
A \code{\link[base:numeric_version]{base::numeric_version()}} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
job$version()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path"></a>}}
\if{latex}{\out{\hypertarget{method-path}{}}}
\subsection{Method \code{path()}}{
Get the paths of file that current \code{EpwSql} uses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$path(type = c("all", "idf", "epw"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{If \code{"all"}, both the \link{Idf} path and \link{Epw} path are
returned. If \code{"idf"}, only IDF path is returned. If \code{"epw"},
only EPW path is returned. If \code{epw} is \code{NULL}, \code{NA} is
returned for EPW path. Default: \code{"all"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$path()} returns the path of IDF or EPW of current job.
}

\subsection{Returns}{
A character vector.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
job$path()
job$path("idf")
job$path("epw")
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Run simulationA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$run(
  epw,
  dir = NULL,
  wait = TRUE,
  force = FALSE,
  echo = wait,
  copy_external = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{epw}}{A path to an \code{.epw} file or an \link{Epw} object. \code{epw} can
also be \code{NULL} which will force design-day-only simulation.
Note this needs at least one \code{Sizing:DesignDay} object exists
in the \code{Idf}. If not given, the \code{epw} input used when creating
this \code{EplusJob} object will be used.}

\item{\code{dir}}{The directory to save the simulation results. If \code{NULL},
the input \code{idf} folder will be used. Default: \code{NULL}.}

\item{\code{wait}}{If \code{TRUE}, R will hang on and wait for the simulation to
complete. EnergyPlus standard output (stdout) and error
(stderr) is printed to R console. If \code{FALSE}, simulation will
be run in a background process.  Default: \code{TRUE}.}

\item{\code{force}}{Only applicable when the last job runs with \code{wait}
equals to \code{FALSE} and is still running. If \code{TRUE}, current
running job is forced to stop and a new one will start.
Default: \code{FALSE}.}

\item{\code{echo}}{Only applicable when \code{wait} is \code{TRUE}. Whether to show
standard output and error from EnergyPlus. Default: same as
\code{wait}.}

\item{\code{copy_external}}{If \code{TRUE}, the external files that current \code{Idf}
object depends on will also be copied into the simulation
output directory. The values of file paths in the Idf will be
changed automatically. Currently, only \code{Schedule:File} class
is supported.  This ensures that the output directory will
have all files needed for the model to run. Default is
\code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$run()} runs the simulation using input IDF and EPW file. If \code{wait}
is \code{FALSE}, the job is run in the background. You can get updated job
status by just
\href{../../eplusr/html/EplusJob.html#method-print}{printing}
the \code{EplusJob} object.

Parameter \code{epw} can be used to reset the EPW file to use for
simulation. If not given, the \code{epw} input used when creating
this \code{EplusJob} object will be used.
}

\subsection{Returns}{
The \code{EplusJob} object itself, invisibly.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# only run design day
job$run(NULL)

# specify output directory
job$run(dir = tempdir())

# run in the background
job$run(wait = TRUE)
# see job status
job$status()

# force to kill background job before running the new one
job$run(force = TRUE)

# do not show anything in the console
job$run(echo = FALSE)

# copy external files used in the model to simulation output directory
job$run(copy_external = TRUE)
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kill"></a>}}
\if{latex}{\out{\hypertarget{method-kill}{}}}
\subsection{Method \code{kill()}}{
Kill current running job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$kill()}\if{html}{\out{</div>}}
}

\subsection{Details}{
\verb{$kill()} kills the background EnergyPlus process if possible. It
only works when simulation runs in non-waiting mode.
}

\subsection{Returns}{
A single logical value of \code{TRUE} or \code{FALSE}, invisibly.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
job$kill()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-status"></a>}}
\if{latex}{\out{\hypertarget{method-status}{}}}
\subsection{Method \code{status()}}{
Get the job status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$status()}\if{html}{\out{</div>}}
}

\subsection{Details}{
\verb{$status()} returns a named list of values that indicates the status of the
job:
\itemize{
\item \code{run_before}: \code{TRUE} if the job has been run before. \code{FALSE} otherwise.
\item \code{alive}: \code{TRUE} if the simulation is still running in the background.
\code{FALSE} otherwise.
\item \code{terminated}: \code{TRUE} if the simulation was terminated during last
simulation. \code{FALSE} otherwise. \code{NA} if the job has not been run yet.
\item \code{successful}: \code{TRUE} if last simulation ended successfully. \code{FALSE}
otherwise. \code{NA} if the job has not been run yet.
\item \code{changed_after}: \code{TRUE} if the IDF file has been changed since last
simulation. \code{FALSE} otherwise. \code{NA} if the job has not been run yet.
}
}

\subsection{Returns}{
A named list of 5 elements.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
job$status()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-errors"></a>}}
\if{latex}{\out{\hypertarget{method-errors}{}}}
\subsection{Method \code{errors()}}{
Read simulation errors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$errors(info = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{info}}{If \code{FALSE}, only warnings and errors are printed.
Default: \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
$errors() returns an \link[=read_err]{ErrFile} object which contains all
contents of the simulation error file (\code{.err}). If \code{info} is \code{FALSE},
only warnings and errors are printed.
}

\subsection{Returns}{
An \link[=read_err]{ErrFile} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
job$errors()

# show all information
job$errors(info = TRUE)
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-output_dir"></a>}}
\if{latex}{\out{\hypertarget{method-output_dir}{}}}
\subsection{Method \code{output_dir()}}{
Get simulation output directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$output_dir(open = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{open}}{If \code{TRUE}, the output directory will be opened.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$output_dir()} returns the output directory of simulation results.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
job$output_dir()

# Below will open output directory
# job$output_dir(open = TRUE)
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-locate_output"></a>}}
\if{latex}{\out{\hypertarget{method-locate_output}{}}}
\subsection{Method \code{locate_output()}}{
Get path of output file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$locate_output(suffix = ".err", strict = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{suffix}}{A string that indicates the file extension of
simulation output. Default: \code{".err"}.}

\item{\code{strict}}{If \code{TRUE}, it will check if the simulation was
terminated, is still running or the file exists or not.
Default: \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$locate_output()} returns the path of a single output file specified
by file suffix.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# get the file path of the error file
job$locate_output(".err", strict = FALSE)

# can use to detect if certain output file exists
job$locate_output(".expidf", strict = TRUE)
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_rdd"></a>}}
\if{latex}{\out{\hypertarget{method-read_rdd}{}}}
\subsection{Method \code{read_rdd()}}{
Read Report Data Dictionary (RDD) file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$read_rdd()}\if{html}{\out{</div>}}
}

\subsection{Details}{
\verb{$read_rdd()} return the core data of Report Data Dictionary (RDD)
file. For details, please see \code{\link[=read_rdd]{read_rdd()}}.
}

\subsection{Returns}{
An \link[=read_rdd]{RddFile} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
job$read_rdd()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_mdd"></a>}}
\if{latex}{\out{\hypertarget{method-read_mdd}{}}}
\subsection{Method \code{read_mdd()}}{
Read Report Data Dictionary (RDD) file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$read_mdd()}\if{html}{\out{</div>}}
}

\subsection{Details}{
\verb{$read_mdd()} return the core data of Meter Data Dictionary (MDD)
file. For details, please see \code{\link[=read_mdd]{read_mdd()}}.
}

\subsection{Returns}{
An \link[=read_mdd]{MddFile} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
job$read_mdd()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_table"></a>}}
\if{latex}{\out{\hypertarget{method-list_table}{}}}
\subsection{Method \code{list_table()}}{
List all table names in EnergyPlus SQL output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$list_table()}\if{html}{\out{</div>}}
}

\subsection{Details}{
\verb{$list_table()} returns all available table and view names in the
EnergyPlus SQLite file.
}

\subsection{Returns}{
A character vector
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
job$list_table()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_table"></a>}}
\if{latex}{\out{\hypertarget{method-read_table}{}}}
\subsection{Method \code{read_table()}}{
Read a single table from EnergyPlus SQL output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$read_table(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A single string specifying the name of table to read.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$read_table()} takes a valid table \code{name} of those from
\href{../../eplusr/html/EplusJob.html#method-list_table}{\code{$list_table()}}
and returns that table data in a \code{\link[data.table:data.table]{data.table::data.table()}} format.
}

\subsection{Returns}{
A \code{\link[data.table:data.table]{data.table::data.table()}}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# read a specific table
job$read_table("Zones")
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-report_data_dict"></a>}}
\if{latex}{\out{\hypertarget{method-report_data_dict}{}}}
\subsection{Method \code{report_data_dict()}}{
Read report data dictionary from EnergyPlus SQL output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$report_data_dict()}\if{html}{\out{</div>}}
}

\subsection{Details}{
\verb{$report_data_dict()} returns a \code{\link[data.table:data.table]{data.table::data.table()}} which
contains all information about report data.

For details on the meaning of each columns, please see "2.20.2.1
ReportDataDictionary Table" in EnergyPlus "Output Details and
Examples" documentation.
}

\subsection{Returns}{
A \code{\link[data.table:data.table]{data.table::data.table()}} of 10 columns:
\itemize{
\item \code{report_data_dictionary_index}: The integer used to link the
dictionary data to the variable data. Mainly useful when joining
different tables
\item \code{is_meter}: Whether report data is a meter data. Possible values:
\code{0} and \code{1}
\item \code{timestep_type}: Type of data timestep. Possible values: \code{Zone} and
\verb{HVAC System}
\item \code{key_value}: Key name of the data
\item \code{name}: Actual report data name
\item \code{reporting_frequency}: Data reporting frequency
\item \code{schedule_name}: Name the the schedule that controls reporting
frequency.
\item \code{units}: The data units
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
job$report_data_dict()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-report_data"></a>}}
\if{latex}{\out{\hypertarget{method-report_data}{}}}
\subsection{Method \code{report_data()}}{
Read report data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$report_data(
  key_value = NULL,
  name = NULL,
  year = NULL,
  tz = "UTC",
  case = "auto",
  all = FALSE,
  wide = FALSE,
  period = NULL,
  month = NULL,
  day = NULL,
  hour = NULL,
  minute = NULL,
  interval = NULL,
  simulation_days = NULL,
  day_type = NULL,
  environment_name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key_value}}{A character vector to identify key values of the
data. If \code{NULL}, all keys of that variable will be returned.
\code{key_value} can also be data.frame that contains \code{key_value}
and \code{name} columns. In this case, \code{name} argument in
\verb{$report_data()} is ignored. All available \code{key_value} for
current simulation output can be obtained using
\href{../../eplusr/html/EplusSql.html#method-report_data_dict}{\code{$report_data_dict()}}.
Default: \code{NULL}.}

\item{\code{name}}{A character vector to identify names of the data. If
\code{NULL}, all names of that variable will be returned. If
\code{key_value} is a data.frame, \code{name} is ignored. All available
\code{name} for current simulation output can be obtained using
\href{../../eplusr/html/EplusSql.html#method-report_data_dict}{\code{$report_data_dict()}}.
Default: \code{NULL}.}

\item{\code{year}}{Year of the date time in column \code{datetime}. If \code{NULL}, it
will calculate a year value that meets the start day of week
restriction for each environment. Default: \code{NULL}.}

\item{\code{tz}}{Time zone of date time in column \code{datetime}. Default:
\code{"UTC"}.}

\item{\code{case}}{If not \code{NULL}, a character column will be added indicates
the case of this simulation. If \code{"auto"}, the name of the IDF
file without extension is used.}

\item{\code{all}}{If \code{TRUE}, extra columns are also included in the returned
\code{\link[data.table:data.table]{data.table::data.table()}}.}

\item{\code{wide}}{If \code{TRUE}, the output is formatted in the same way as
standard EnergyPlus csv output file.}

\item{\code{period}}{A Date or POSIXt vector used to specify which time
period to return. The year value does not matter and only
month, day, hour and minute value will be used when
subsetting. If \code{NULL}, all time period of data is returned.
Default: \code{NULL}.}

\item{\code{month, day, hour, minute}}{Each is an integer vector for month,
day, hour, minute subsetting of \code{datetime} column when
querying on the SQL database. If \code{NULL}, no subsetting is
performed on those components. All possible \code{month}, \code{day},
\code{hour} and \code{minute} can be obtained using
\href{../../eplusr/html/EplusSql.html#method-read_table}{\code{$read_table("Time")}}.
Default: \code{NULL}.}

\item{\code{interval}}{An integer vector used to specify which interval
length of report to extract. If \code{NULL}, all interval will be
used. Default: \code{NULL}.}

\item{\code{simulation_days}}{An integer vector to specify which simulation
day data to extract. Note that this number resets after warmup
and at the beginning of an environment period. All possible
\code{simulation_days} can be obtained using
\href{../../eplusr/html/EplusSql.html#method-read_table}{\code{$read_table("Time")}}.
If \code{NULL}, all simulation days will be used. Default: \code{NULL}.}

\item{\code{day_type}}{A character vector to specify which day type of data
to extract. All possible day types are: \code{Sunday}, \code{Monday},
\code{Tuesday}, \code{Wednesday}, \code{Thursday}, \code{Friday}, \code{Saturday},
\code{Holiday}, \code{SummerDesignDay}, \code{WinterDesignDay}, \code{CustomDay1},
and \code{CustomDay2}. All possible values for current simulation
output can be obtained using
\href{../../eplusr/html/EplusSql.html#method-read_table}{\code{$read_table("Time")}}.}

\item{\code{environment_name}}{A character vector to specify which
environment data to extract. If \code{NULL}, all environment data
are returned. Default: \code{NULL}. All possible
\code{environment_name} for current simulation output can be
obtained using:\preformatted{$read_table("EnvironmentPeriods")
}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$report_data()} extracts the report data in a
\code{\link[data.table:data.table]{data.table::data.table()}} using key values, variable names and other
specifications.

\verb{$report_data()} can also directly take all or subset output from
\verb{$report_data_dict()} as input, and extract all data specified.

The returned column numbers varies depending on \code{all} argument.
\itemize{
\item \code{all} is \code{FALSE}, the returned \code{\link[data.table:data.table]{data.table::data.table()}} has 6 columns:
\itemize{
\item \code{case}: Simulation case specified using \code{case} argument
\item \code{datetime}: The date time of simulation result
\item \code{key_value}: Key name of the data
\item \code{name}: Actual report data name
\item \code{units}: The data units
\item \code{value}: The data value
}
\item \code{all} is \code{TRUE}, besides columns described above, extra columns are also
included:
\itemize{
\item \code{month}: The month of reported date time
\item \code{day}: The day of month of reported date time
\item \code{hour}: The hour of reported date time
\item \code{minute}: The minute of reported date time
\item \code{dst}: Daylight saving time indicator. Possible values: \code{0} and \code{1}
\item \code{interval}: Length of reporting interval
\item \code{simulation_days}: Day of simulation
\item \code{day_type}: The type of day, e.g. \code{Monday}, \code{Tuesday} and etc.
\item \code{environment_period_index}: The indices of environment.
\item \code{environment_name}: A text string identifying the environment.
\item \code{is_meter}: Whether report data is a meter data. Possible values: \code{0} and
\code{1}
\item \code{type}: Nature of data type with respect to state. Possible values: \code{Sum}
and \code{Avg}
\item \code{index_group}: The report group, e.g. \code{Zone}, \code{System}
\item \code{timestep_type}: Type of data timestep. Possible values: \code{Zone} and \verb{HVAC System}
\item \code{reporting_frequency}: The reporting frequency of the variable, e.g.
\verb{HVAC System Timestep}, \verb{Zone Timestep}.
\item \code{schedule_name}: Name of the the schedule that controls reporting
frequency.
}
}

With the \code{datetime} column, it is quite straightforward to apply time-series
analysis on the simulation output. However, another painful thing is that
every simulation run period has its own \verb{Day of Week for Start Day}. Randomly
setting the \code{year} may result in a date time series that does not have
the same start day of week as specified in the RunPeriod objects.

eplusr provides a simple solution for this. By setting \code{year} to \code{NULL},
which is the default behavior, eplusr will calculate a year value (from
current year backwards) for each run period that compliances with the start
day of week restriction.

It is worth noting that EnergyPlus uses 24-hour clock system where 24 is only
used to denote midnight at the end of a calendar day. In EnergyPlus output,
"00:24:00" with a time interval being 15 mins represents a time period from
"00:23:45" to "00:24:00", and similarly "00:15:00" represents a time period
from "00:24:00" to "00:15:00" of the next day. This means that if current day
is Friday, day of week rule applied in schedule time period "00:23:45" to
"00:24:00" (presented as "00:24:00" in the output) is also Friday, but not
Saturday. However, if you try to get the day of week of time "00:24:00" in R,
you will get Saturday, but not Friday. This introduces inconsistency and may
cause problems when doing data analysis considering day of week value.

With \code{wide} equals \code{TRUE}, \verb{$report_data()} will format the simulation output
in the same way as standard EnergyPlus csv output file. Sometimes this can be
useful as there may be existing tools/workflows that depend on this format.
When both \code{wide} and \code{all} are \code{TRUE}, columns of runperiod environment names
and date time components are also returned, including:
\verb{environment_period_index", "environment_name}, \code{simulation_days},
\code{datetime}, \code{month}, \code{day}, \code{hour}, \code{minute}, \code{day_type}.

For convenience, input character arguments matching in
\verb{$report_data()} are \strong{case-insensitive}.
}

\subsection{Returns}{
A \code{\link[data.table:data.table]{data.table::data.table()}}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# read all report data
job$report_data()

# specify output variables using report data dictionary
dict <- job$report_data_dict()
job$report_data(dict[units == "C"])

# specify output variables using 'key_value' and 'name'
job$report_data("environment", "site outdoor air drybulb temperature")

# explicitly specify year value and time zone
job$report_data(dict[1], year = 2020, tz = "Etc/GMT+8")

# explicitly specify case name
job$report_data(dict[1], case = "example")

# get all possible columns
job$report_data(dict[1], all = TRUE)

# return in a format that is similar as EnergyPlus CSV output
job$report_data(dict[1], wide = TRUE)

# return in a format that is similar as EnergyPlus CSV output with
# extra columns
job$report_data(dict[1], wide = TRUE, all = TRUE)

# only get data at the working hour on the first Monday
job$report_data(dict[1], hour = 8:18, day_type = "monday", simulation_days = 1:7)

# only get specified run period data
job$read_table("EnvironmentPeriods") # possible environment name
job$report_data(dict[1], environment_name = "San Francisco Intl Ap CA USA TMY3 WMO#=724940")
# can also be done using 'environment_period_index' column
job$report_data(dict[1], all = TRUE)[environment_period_index == 3L]
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tabular_data"></a>}}
\if{latex}{\out{\hypertarget{method-tabular_data}{}}}
\subsection{Method \code{tabular_data()}}{
Read tabular data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$tabular_data(
  report_name = NULL,
  report_for = NULL,
  table_name = NULL,
  column_name = NULL,
  row_name = NULL,
  wide = FALSE,
  string_value = !wide
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{report_name, report_for, table_name, column_name, row_name}}{Each is
a character vector for subsetting when querying the SQL
database.  For the meaning of each argument, please see the
description above.}

\item{\code{wide}}{If \code{TRUE}, each table will be converted into the similar
format as it is shown in EnergyPlus HTML output file. Default:
\code{FALSE}.}

\item{\code{string_value}}{Only applicable when \code{wide} is \code{TRUE}. If
\code{string_value} is \code{FALSE}, instead of keeping all values as
characters, values in possible numeric columns are converted
into numbers. Default: the opposite of \code{wide}. Possible
numeric columns indicate column that:
\itemize{
\item columns that have associated units
\item columns that contents numbers
}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\verb{$tabular_data()} extracts the tabular data in a
\code{\link[data.table:data.table]{data.table::data.table()}} using report, table, column and row name
specifications. The returned \code{\link[data.table:data.table]{data.table::data.table()}} has
9 columns:
\itemize{
\item \code{case}: Simulation case specified using \code{case} argument
\item \code{index}: Tabular data index
\item \code{report_name}: The name of the report that the record belongs to
\item \code{report_for}: The \code{For} text that is associated with the record
\item \code{table_name}: The name of the table that the record belongs to
\item \code{column_name}: The name of the column that the record belongs to
\item \code{row_name}: The name of the row that the record belongs to
\item \code{units}: The units of the record
\item \code{value}: The value of the record \strong{in string format} by default
}

For convenience, input character arguments matching in
\verb{$tabular_data()} are \strong{case-insensitive}.
}

\subsection{Returns}{
A \code{\link[data.table:data.table]{data.table::data.table()}} with 8 columns (when \code{wide} is
\code{FALSE}) or a named list of \code{\link[data.table:data.table]{data.table::data.table()}}s where the
names are the combination of \code{report_name}, \code{report_for} and
\code{table_name}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# read all tabular data
job$tabular_data()

# explicitly specify data you want
str(job$tabular_data(
    report_name = "AnnualBuildingUtilityPerformanceSummary",
    table_name = "Site and Source Energy",
    column_name = "Total Energy",
    row_name = "Total Site Energy"
))

# get tabular data in wide format and coerce numeric values
str(job$tabular_data(
    report_name = "AnnualBuildingUtilityPerformanceSummary",
    table_name = "Site and Source Energy",
    column_name = "Total Energy",
    row_name = "Total Site Energy",
    wide = TRUE, string_value = FALSE
))
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print \code{EplusSql} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EplusJob$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
\verb{$print()} shows the core information of this \code{EplusJob} object,
including the path of model and weather, the version and path of
EnergyPlus used to run simulations, and the simulation job status.

\verb{$print()} is quite useful to get the simulation status, especially
when \code{wait} is \code{FALSE} in \verb{$run()}. The job status will be updated
and printed whenever \verb{$print()} is called.
}

\subsection{Returns}{
The \code{EplusSql} object itself, invisibly.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
job$print()
}

}
\if{html}{\out{</div>}}

}

}
}
