% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplusr.R
\docType{package}
\name{eplusr-package}
\alias{tibble-package}
\alias{eplusr-package}
\title{eplusr: A Toolkit for Using EnergyPlus in R}
\description{
This package provides tools to run EnergyPlus and collect,
analyse, summary and plot parametric simulation results in R.

Display options for \code{tbl_df}, used by \code{\link[=trunc_mat]{trunc_mat()}} and
(indirectly) by \code{\link[=print.tbl]{print.tbl()}}.
}
\details{
EnergyPlus has been widely used for XXXX. \code{eplusr}
}
\section{Methods}{


\code{tbl_df} implements four important base methods:

\describe{
\item{print}{By default only prints the first 10 rows (at most 20), and the
columns that fit on screen; see \code{\link[=print.tbl]{print.tbl()}}}
\item{\code{[}}{Does not simplify (drop) by default, returns a data frame}
\item{\code{[[}, \code{$}}{Calls \code{\link[=.subset2]{.subset2()}} directly,
so is considerably faster. Returns \code{NULL} if column does not exist,
\code{$} warns.}
}
}

\section{Important functions}{


\code{\link[=tibble]{tibble()}} and \code{\link[=tribble]{tribble()}} for construction,
\code{\link[=as_tibble]{as_tibble()}} for coercion,
and \code{\link[=print.tbl]{print.tbl()}} and \code{\link[=glimpse]{glimpse()}} for display.
}

\section{Package options}{

\itemize{
\item \code{tibble.print_max}: Row number threshold: Maximum number of rows
printed. Set to \code{Inf} to always print all rows.  Default: 20.
\item \code{tibble.print_min}: Number of rows printed if row number
threshold is exceeded. Default: 10.
\item \code{tibble.width}: Output width. Default: \code{NULL} (use
\code{width} option).
\item \code{tibble.max_extra_cols}: Number of extra columns
printed in reduced form. Default: 100.
}
}

\examples{
tibble(a = 1:26, b = letters)
as_tibble(iris)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hongyuanjia/eplusr}
  \item Report bugs at \url{https://github.com/hongyuanjia/eplusr/issues}
}

}
\author{
\strong{Maintainer}: Hongyuan Jia \email{hongyuanjia@cqu.edu.cn}

}
