% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplusr.R
\docType{package}
\name{eplusr-package}
\alias{tibble-package}
\alias{eplusr-package}
\alias{eplusr_opt}
\title{eplusr: A Toolkit for Using EnergyPlus in R}
\usage{
eplusr_opt(x)
}
\description{
This package provides tools to run EnergyPlus and collect,
analyse, summary and plot parametric simulation results in R.

Display options for \code{tbl_df}, used by \code{\link[=trunc_mat]{trunc_mat()}} and
(indirectly) by \code{\link[=print.tbl]{print.tbl()}}.
}
\details{
EnergyPlus has been widely used for XXXX. \code{eplusr}
}
\section{Methods}{


\code{tbl_df} implements four important base methods:

\describe{
\item{print}{By default only prints the first 10 rows (at most 20), and the
columns that fit on screen; see \code{\link[=print.tbl]{print.tbl()}}}
\item{\code{[}}{Does not simplify (drop) by default, returns a data frame}
\item{\code{[[}, \code{$}}{Calls \code{\link[=.subset2]{.subset2()}} directly,
so is considerably faster. Returns \code{NULL} if column does not exist,
\code{$} warns.}
}
}

\section{Important functions}{


\code{\link[=tibble]{tibble()}} and \code{\link[=tribble]{tribble()}} for construction,
\code{\link[=as_tibble]{as_tibble()}} for coercion,
and \code{\link[=print.tbl]{print.tbl()}} and \code{\link[=glimpse]{glimpse()}} for display.
}

\section{Package options}{

}

\examples{
tibble(a = 1:26, b = letters)
as_tibble(iris)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hongyuanjia/eplusr}
  \item Report bugs at \url{https://github.com/hongyuanjia/eplusr/issues}
}

}
\author{
\strong{Maintainer}: Hongyuan Jia \email{hongyuanjia@cqu.edu.cn} (0000-0002-0075-8183)

}
