% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idd.R
\docType{class}
\name{idd}
\alias{idd}
\title{Parse EnergyPlus IDD files}
\value{
For \code{use_idd}, an Idd object.
}
\description{
eplusr provides parsing of and programmatic access to EnergyPlus
Input Data Dictionary (IDD) files, and objects. It contains all data needed
to parse EnergyPlus models. \code{Idd} class provides parsing and printing while
\code{IddObject} provides detailed information of curtain class.
}
\details{
The \code{Idd} objects for EnergyPlus 8.5 to 8.8 have been pre-parsed and
stored internally and will automatically be used when parsing IDF files and
strings.  Internally, the powerful \code{data.table} package is used to speed
up the whole process and store the results. However, it will still take about
5-6 sec to parse an IDD file.

Normally, you may not need to parse any Energy+.idd file unless your model
is produced by EnergyPlus whose version is lower than 8.5. If so, it is
suggested to store the parsed IDD object and directly pass it to the \code{idd}
argument in \code{\link[=read_idf]{read_idf()}} in order to avoid the parsing process whenever you
read a model of that version.

Under the hook, eplusr uses a SQL-like structure to store both IDF and IDD
data in \code{data.frame} format. Every IDD will be parsed and stored in four
tables:
\itemize{
\item \code{group}: contains group index and group names.
\item \code{class}: contains class names and properties.
\item \code{class_reference}: contains reference names of classes.
\item \code{field}: contains field names and field properties.
\item \code{field_reference}: contains reference names of fields.
\item \code{field_default}: contains default values of fields.
\item \code{field_choice}: contains choices of choice-type fields.
\item \code{field_range}: contains range data of fields.
\item \code{field_object_list}: contains object-list data of fields.
\item \code{field_external_list}: contains external-list data of fields.
}
}
\section{Usage}{
\preformatted{# read
idd <- use_idd(idd)

# basic info
idd$version()
idd$build()

idd$group_index(group = NULL)
idd$group_name()
idd$is_valid_group(group)
idd$from_group(class)

idd$class_index(class = NULL)
idd$class_name()
idd$is_valid_class(class)

idd$required_class_name()
idd$unique_class_name()
idd$extenesible_class_name()

# idd object
idd$object(class)
idd$object_in_group(group)

idd$print()
print(idd)
}
}

\section{Arguments}{

\itemize{
\item \code{idd}: Path to an EnergyPlus Input Data Dictionary (IDD) file, usually
named as \code{Energy+.idd} or a valid version of pre-parsed IDD (8.3 - 8.9).
\item \code{group}: A valid group name or valid group names.
\item \code{class}: A valid class name or valid class names.
}
}

\section{Detail}{


\code{use_idd} will parses an EnergyPlus Input Data Dictionary (IDD) file, and
returns an \code{Idd} object. If \code{idd} is a valid version of pre-parsed IDD file,
then the pre-parsed \code{Idd} object will be returned.

\code{$version} returns the version string.

\code{$build} returns the build tag string.

\code{$group_index} returns integer indexes (indexes of name appearance in
the IDD file) of specified groups.

\code{$group_name} returns all group names.

\code{$from_group} returns the names of group that specified classes belongs to.

\code{$is_valid_group} return \code{TRUE} if the input is a valid group name.

\code{$class_index} returns integer indexes (indexes of name appearance in
the IDD file) of specified classes.

\code{$class_name} returns all class names.

\code{$required_class_name} returns the names of all required classes.

\code{$unique_class_name} returns the names of all unique classes.

\code{$extensible_class_name} returns the names of all extensible classes.

\code{$is_valid_class} return \code{TRUE} if the input is a valid class name.

\code{$object} returns a list of \code{IddObject}s of specified classes.

\code{$object_in_group} returns a list of \code{IddObject}s in that group.

For details about \code{IddObject}, please see \link{idd_object}.
}

\references{
\href{https://github.com/NREL/EnergyPlus/tree/develop/src/IDF_Editor}{IDFEditor
source code}
}
\author{
Hongyuan Jia
}
