% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_object.R
\docType{class}
\name{IDFObject}
\alias{IDFObject}
\title{EnergyPlus IDF objects}
\value{
An IDFObject object
}
\description{
\code{IDFObject} is a R6 class used internally as members in \code{IDF} R6
class. \code{IDFObject} inherits \code{\link{IDDObject}}, so all methods of
\code{IDDObject} are available for \code{IDFobject}.
}
\section{Usage}{
\preformatted{
idfobj <- IDFObject$new(list, idd)

idfobj$get_comment()
idfobj$set_comment(comment, append = TRUE, wrap = 0L)
idfobj$get_value(index = NULL, name = NULL)
idfobj$set_value(...)

idfobj$check()

idfobj$out(comment = TRUE, leading = 4L, ip_units = FALSE, sep_at = 29L,
           add_blank = FALSE)
idfobj$out_lines(index = NULL, ip_units = FALSE)

idfobj$print(comment = TRUE)

print(iddobj)
}

For other methods inherited from \code{IDDObject}, please see
\code{\link{IDDObject}}.
}

\section{Detail}{


\subsection{Create}{\preformatted{idfobj <- IDFObject$new(list, idd)
}

\code{IDFObject$new(list, idd)} creates an IDFObject using parsed IDF object data
stored in a list and an IDD object.

\strong{Arguments}:
\itemize{
\item \code{list}: A list. The list contains three named elements:
\code{object_id}, \code{value} and \code{comment}, where:
\itemize{
\item \code{object_id}: a unique integer to reference this IDFObject.
\item \code{value}: a list contains all values of this IDFObject. All values can
be totally strings, but they will be converted to corresponding types
during initialization.
\item \code{comment}: a character vector. Elements will be put line by line when
saved to \code{*.idf} files.
}
\item \code{idd}: An IDD object created by \code{IDD$new()}.
}
}

\subsection{Value}{\preformatted{idfobj$get_value(index = NULL, name = NULL)
}

\code{$get_value(index, name)} return values of certain fields specified by
\code{index} or \code{name}. The values are returned as a list and returned invisibly.

\code{$set_value(...)} sets values given in \code{...} to fields.
}
}

\seealso{
\code{\link{IDDObject}}
}
\author{
Hongyuan Jia
}
