% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idd.R
\name{parse_idd}
\alias{parse_idd}
\title{Parse Energy+.idd file}
\usage{
parse_idd(path)
}
\arguments{
\item{path}{Path to 'Energy+.idd' file}
}
\value{
A list contains the IDD version, build, parsed class data and parsed
field data. Both class data and field data are stored in data.tables.
}
\description{
Parse Energy+.idd file
}
\details{
The IDD files for EnergyPlus 8.0 to 8.8 have been pre-parsed and
stored internally and will automatically be used when parsing \code{IDF} and
\code{IMF} files. The parsing process was basically the same as that was
implemented in IDFEditor distributed with EnergyPlus, but using the powerful
\code{data.table} package to speed up the whole process and store the
results. The souce codes of IDFEditor can be found on
\href{GitHub}{https://github.com/NREL/EnergyPlus/tree/develop/src/IDF_Editor}:
. However, it will still take about 3-4 sec to parse an IDD file which is
much slower than IDFEditor written in Visual Basic.
}
