% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.R
\name{parse_idf}
\alias{parse_idf}
\title{Parse EnergyPlus models}
\usage{
parse_idf(filepath, idd = NULL)
}
\arguments{
\item{idd}{Path to \code{Energy+.idd} or an IDD object.}

\item{path}{Path to EnergyPlus \code{IDF} or \code{IMF} file. The file
extension does not matter. So models stored in \code{TXT} file are still able
to correctly be parsed.}
}
\value{
A list contains the IDF version, option data, class data, value data,
comment data and field reference data.
}
\description{
Parse EnergyPlus models
}
