% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl-idf.R
\name{standardize_idf_value}
\alias{standardize_idf_value}
\title{Standardize Value Data}
\usage{
standardize_idf_value(
  idd_env,
  idf_env,
  dt_value,
  type = c("choice", "reference"),
  keep = TRUE
)
}
\arguments{
\item{idd_env}{An environment or list contains IDD tables including class,
field, and reference.}

\item{idf_env}{An environment or list contains IDF tables including object,
value, and reference.}

\item{dt_value}{An data.table of object field values.}

\item{type}{A character vector to specify what type of values to be
standardized. Should be a subset of \code{c("choice", "reference")}.
Default: \code{c("choice", "reference")}.}

\item{keep}{If \code{TRUE}, the original value will be kept even if it is invalid.
If \code{FALSE}, invalid values will be converted into NAs. Default: \code{TRUE}.}
}
\value{
A data.table
}
\description{
Standardize Value Data
}
\keyword{internal}
