% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idd.R
\name{use_idd}
\alias{use_idd}
\title{Use a specific EnergyPlud Input Data Dictionary (IDD) file}
\usage{
use_idd(idd)
}
\arguments{
\item{idd}{Path to an EnergyPlus Input Data Dictionary (IDD) file, usually
named as \code{Energy+.idd} or a valid version of pre-parsed IDD (8.3 - 8.9).}
}
\value{
An \code{Idd} object
}
\description{
\code{use_idd} takes a path of an EnergyPlus Input Data Dictionary (IDD) file,
usually named "Energy+.idd" and return an \code{Idd} object. For details on \code{Idd}
class, please see \link{idd}.
}
